% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{ancestor}
\alias{ancestor}
\title{Routine function to get ancestors of a subject}
\usage{
ancestor(idx, momx, dadx)
}
\arguments{
\item{idx}{subject index}

\item{momx}{vector of mother indexs}

\item{dadx}{vector of father indexs}
}
\value{
a vector of ancestor index
}
\description{
For each spouse pair, find out if it should be
connected with a double line.  This is the case if they have
a common ancestor
}
