% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\docType{methods}
\name{as.data.frame,Pedigree-method}
\alias{as.data.frame,Pedigree-method}
\title{Convert a Pedigree object to a data.frame}
\usage{
\S4method{as.data.frame}{Pedigree}(x)
}
\arguments{
\item{x}{A Pedigree object.}
}
\value{
A data.frame with the individuals informations.
}
\description{
Convert a Pedigree object to a data.frame
}
