% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.pedigree.R
\name{as.data.frame.pedigree}
\alias{as.data.frame.pedigree}
\title{data.frame from a pedigree object}
\usage{
\method{as.data.frame}{pedigree}(x, ...)
}
\arguments{
\item{x}{Pedigree object}

\item{...}{Additional arguments passed to internal methods}
}
\value{
A data.frame with the data necessary to re-create the pedigree,
minus special relationships.
}
\description{
Extract the internal data from a pedigree object into a data.frame
}
\examples{
data(sample.ped)
ped <- with(sample.ped,pedigree(id, father, mother, sex, affected))
as.data.frame.pedigree(ped)

}
\seealso{
\code{\link{pedigree}}
}
\author{
Jason Sinnwell
}
