% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.pedigreeList.R
\name{as.data.frame.pedigreeList}
\alias{as.data.frame.pedigreeList}
\title{data.frame from a pedigreeList object}
\usage{
\method{as.data.frame}{pedigreeList}(x, ...)
}
\arguments{
\item{x}{Pedigree object}

\item{...}{Additional arguments passed to internal methods}
}
\value{
A data.frame with the data necessary to re-create the pedigree,
minus special relationships.
}
\description{
Extract the internal data from a pedigree object into a data.frame
}
\examples{
data(sample.ped)
ped <- with(sample.ped, pedigree(id, father, mother,
 sex, affected, famid = ped))
as.data.frame(ped)

}
\seealso{
`pedigree`
}
\author{
Louis Le Nézet
}
