% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\docType{methods}
\name{as.list,Pedigree-method}
\alias{as.list,Pedigree-method}
\title{Convert a Pedigree object to a list}
\usage{
\S4method{as.list}{Pedigree}(x)
}
\arguments{
\item{x}{A Pedigree object.}
}
\value{
A list with all the slots of the Pedigree object.
}
\description{
Convert a Pedigree object to a list
}
