% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_hint.R
\name{best_hint}
\alias{best_hint}
\title{Best hint for alignement}
\usage{
best_hint(obj, ...)
}
\arguments{
\item{ped}{A pedigree object}

\item{wt}{A vector of weights for the three error measures}

\item{tolerance}{The maximum stress level to accept}
}
\value{
Hint object
}
\description{
When computer time is cheap, use this routine to get a 'best' pedigree.
}
\details{
The auto_hint routine will rearrange sibling order, but not founder order.
This calls auto_hint with every possible founder order, and finds that
plot with the least 'stress':
       wt[1] * number of dotted arcs
     + wt[2] * lengths of dotted arcs
     + wt[3] * lengths of parent-child
bends perfect: if anything meets this stress level, keep it
}
\seealso{
`auto_hint`
}
