% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besthint.R
\name{besthint}
\alias{besthint}
\title{Best hint for alignement}
\usage{
besthint(ped, wt = c(1000, 10, 1), tolerance = 0)
}
\arguments{
\item{tolerance}{}
}
\description{
When computer time is cheap, use this routine to get a 'best' pedigree.
}
\details{
The autohint routine will rearrange sibling order, but not founder order.
This calls autohint with every possible founder order, and finds that
plot with the least 'stress':
       wt[1] * number of dotted arcs
     + wt[2] * lengths of dotted arcs
     + wt[3] * lengths of parent-child
bends perfect: if anything meets this stress level, keep it
}
\seealso{
`autohint`
}
