% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_size.R
\docType{methods}
\name{bit_size}
\alias{bit_size}
\alias{bit_size,character-method}
\alias{bit_size,character}
\alias{bit_size,Pedigree-method}
\alias{bit_size,Pedigree}
\title{Get pedigree bit_size}
\usage{
bit_size(obj, ...)

## S4 method for signature 'character'

bit_size(dadid, momid, missid = "0")

## S4 method for signature 'Pedigree'

bit_size(ped, missid = "0")
}
\arguments{
\item{...}{Additional arguments passed to methods}

\item{momid}{Vector of mothers ids}

\item{missid}{Character defining the missing ids}

\item{dadid}{Vector of fathers ids}

\item{ped}{A pedigree object}
}
\value{
A list with the following components:
\itemize{
\item bit_size The bit_size of input pedigree
\item nFounder The number of founders in the pedigree
\item nNonFounder The number of nonfounders in the pedgiree
}
}
\description{
Calculate pedigree bit_size, defined as :

\eqn{
  2 \times NbNonFounders - NbFounders
}
}
\details{
This is a utility function used in \code{shrink()}
to calculate the bit_size of a pedigree.
}
\seealso{
\code{\link[=shrink]{shrink()}}
}
