% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitSize.R
\name{bitSize}
\alias{bitSize}
\title{Get pedigree bitsize}
\usage{
bitSize(obj, ...)
}
\arguments{
\item{dadid}{Vector of fathers ids}

\item{momid}{Vector of mothers ids}

\item{missid}{Character defining the missing ids}
}
\value{
A list with the following components:
## bitSize
The bitSize of input pedigree
## nFounder
The number of founders in the pedigree
## nNonFounder
The number of nonfounders in the pedgiree
}
\description{
Calculate pedigree bitsize, defined as 2 * # NonFounders - # Founders
}
\details{
This is a utility function used in `pedigree.shrink()`
}
\seealso{
`pedigree.shrink`
}
