% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{c,Ped-method}
\alias{c,Ped-method}
\title{Bind two Ped objects}
\usage{
\S4method{c}{Ped}(x, y)
}
\arguments{
\item{x}{A Ped object.}

\item{y}{A Ped object.}
}
\value{
A Ped object containing the individuals of x and y.
}
\description{
Bind two Ped objects by row.
The metadata need to be the same.
}
