% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{c,Rel-method}
\alias{c,Rel-method}
\title{Bind two Rel objects}
\usage{
\S4method{c}{Rel}(x, y)
}
\arguments{
\item{x}{A Rel object.}

\item{y}{A Rel object.}
}
\value{
A Rel object containing the relationships of x and y.
}
\description{
Bind two Rel objects by row.
The metadata need to be the same.
}
