% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\name{$,Pedigree-method}
\alias{$,Pedigree-method}
\title{Extract parts of a Pedigree object}
\usage{
\S4method{$}{Pedigree}(x, name)
}
\arguments{
\item{x}{A Pedigree object.}

\item{name}{A slot name.}
}
\value{
The slot \code{name} present in the pedigree object.
}
\description{
Extract parts of a Pedigree object
}
