% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\name{$<-,Pedigree,ANY-method}
\alias{$<-,Pedigree,ANY-method}
\title{Replace parts of a Pedigree object}
\usage{
\S4method{$}{Pedigree,ANY}(x, name) <- value
}
\arguments{
\item{x}{A Pedigree object.}

\item{name}{A slot name.}

\item{value}{A vector of values to replace.}
}
\value{
The pedigree object with the slot \code{name} replaced by \code{value}.
}
\description{
Replace parts of a Pedigree object
}
