% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindepth.R
\name{chaseup}
\alias{chaseup}
\title{Chase up the ancestors of a subject}
\usage{
chaseup(x, midx, didx)
}
\arguments{
\item{x}{a vector of subject ids}

\item{midx}{a vector of mother ids}

\item{didx}{a vector of father ids}
}
\value{
a vector of all ancestors subjects connected to x
}
\description{
Chase up the ancestors of a subject
}
\examples{
chaseup(1, c(3,0,0), c(2,0,0))
chaseup(1, c(3,4,0,0,0), c(2,5,0,0,0))
}
\keyword{internal,}
\keyword{kindepth}
