% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check for columns name usage}
\usage{
check_columns(df, cols_needed, cols_used, cols_to_use, others_cols = FALSE)
}
\arguments{
\item{df}{The dataframe to use}

\item{cols_needed}{A vector of columns needed}

\item{cols_used}{A vector of columns that are used by the script and
that will be overwritten.}

\item{cols_to_use}{A vector of optional columns that are authorized.}

\item{others_cols}{Boolean defining if non defined columns should be allowed.}
}
\value{
Dataframe with only the column allowed and all the column
to be used by the script initialised to NA.
}
\description{
Check for presence / absence of columns names
depending on their need
}
\details{
3 types of columns are here checked:
- 1 `cols_needed` those columns need to be present if any is missing
an error will be prompted and the script will stop
- 2 `cols_used` those columns will be used in the script and will be
overwritten to NA.
- 3 `cols_to_use` those columns are optional and will be recognise
if present.
}
\examples{
data.frame
df <- data.frame(ColN1 = c(1, 2), ColN2 = 4,
         ColU1 = "B", ColU2 = "1",
         ColTU1 = "A", ColTU2 = 3,
         ColNR1 = 4, ColNR2 = 5)
check_columns(df, c("ColN1", "ColN2"), c("ColU1", "ColU2"),
  c("ColTU1", "ColTU2"))

}
