% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hints.R
\name{check_hints}
\alias{check_hints}
\title{Detect hints inconsistencies}
\usage{
check_hints(hints, sex)
}
\arguments{
\item{hints}{A list of hints}

\item{sex}{A vector with the sex of all the individuals}
}
\description{
This routine tries to detect inconsistencies in spousal hints.
}
\details{
These arise in auto_hint with complex pedigrees.
One can have ABA (subject A is on both the left and the right of B),
cycles, etc.
Actually, these used to arise in auto_hint.
Users can introduce problems as well if they modify the hints.
}
\seealso{
\code{auto_hint}, \code{best_hint}
}
