% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_num_na}
\alias{check_num_na}
\title{Is numeric or NA}
\usage{
check_num_na(var, na_as_num = TRUE)
}
\arguments{
\item{var}{Vector of value to test}

\item{na_as_num}{Boolean defining if the \code{NA} string should be
considered as numerical values}
}
\value{
A vector of boolean of the same size as \strong{var}
}
\description{
Check if a variable given is numeric or NA
}
\details{
Check if the values in \strong{var} are numeric or if they are
\code{NA} in the case that \code{na_as_num} is set to TRUE.
}
\keyword{internal}
