% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{check_slot_fd}
\alias{check_slot_fd}
\title{Check if the fields are present in an object slot}
\usage{
check_slot_fd(object, slot = NULL, fields = character())
}
\arguments{
\item{object}{An object.}

\item{slot}{A slot of object.}

\item{fields}{A character vector with the fields to check.}
}
\value{
A character vector with the errors if any.
}
\description{
Check if the fields are present in an object slot
}
