% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{check_values}
\alias{check_values}
\title{Check values in a slot}
\usage{
check_values(val, ref, present = TRUE)
}
\arguments{
\item{present}{A logical value indicating if the values should be present
or not}

\item{object}{An object.}

\item{slot}{A slot of the object.}

\item{column}{A column of the slot.}

\item{values}{A vector of values to check.}
}
\value{
A character vector with the errors if any.
}
\description{
Check if the all the values in a slot are in a vector of values.
}
