% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{checkColnamesSlot}
\alias{checkColnamesSlot}
\title{Check if the colnames are present in an object slot}
\usage{
checkColnamesSlot(object, slot = NULL, colnames = character())
}
\arguments{
\item{object}{An object.}

\item{slot}{A slot of object.}

\item{colnames}{A character vector with the colnames to check.}
}
\value{
A character vector with the errors if any.
}
\description{
Check if the colnames are present in an object slot
}
