% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHint.R
\name{checkHint}
\alias{checkHint}
\title{Remove inconsistencies}
\usage{
checkHint(hints, sex)
}
\arguments{
\item{sex}{}
}
\description{
This routine tries to remove inconsistencies in spousal hints.
}
\details{
These arise in autohint with complex pedigrees.
One can have ABA (subject A is on both the left and the right of B),
cycles, etc.
Actually, these used to arise in autohint.
Users can introduce problems as well if they modify the hints.
}
\seealso{
`autohint`, `besthint`
}
