% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{checkValuesSlot}
\alias{checkValuesSlot}
\title{Check values in a slot}
\usage{
checkValuesSlot(object, slot, column, values, present = TRUE)
}
\arguments{
\item{object}{An object.}

\item{slot}{A slot of the object.}

\item{column}{A column of the slot.}

\item{values}{A vector of values to check.}

\item{present}{A logical value indicating if the values should be present
or not}
}
\value{
A character vector with the errors if any.
}
\description{
Check if the all the values in a slot are in a vector of values.
}
