% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{circfun}
\alias{circfun}
\title{Generate a circular element}
\usage{
circfun(nslice, n = 50)
}
\arguments{
\item{nslice}{number of slices in the circle}

\item{n}{Total number of points in the circle}
}
\value{
a list of x and y coordinates
}
\description{
Generate a circular element
}
\keyword{internal}
