% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendant.pedigree.R
\name{descendant.pedigree}
\alias{descendant.pedigree}
\title{Find all the descendants}
\usage{
descendant.pedigree(idlist, id, dadid, momid)
}
\arguments{
\item{momid}{}
}
\description{
Find all the descendants of a particular person given a pedigree or a list
}
