% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachPackage.R
\name{detachPackage}
\alias{detachPackage}
\title{detach a loaded library if its loaded}
\usage{
detachPackage(pkg, character.only = FALSE)
}
\arguments{
\item{pkg}{The package name}

\item{charactr.only}{Somethingelse}
}
\description{
detach a loaded library if its loaded
}
