% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_cont_table}
\alias{df_cont_table}
\title{Data frame to contingency table}
\usage{
df_cont_table(df, var1, threshold1 = NULL, var2 = NULL, threshold2 = NULL)
}
\arguments{
\item{df}{Dataframe containing the family informations}

\item{var1, var2}{First and second column variable to select}

\item{threshold1, threshold2}{Thresholds to apply to \code{var1} and \code{var2}}
}
\description{
Summarise two variables in a contingency table
}
\details{
Proccess two variable from a dataframe in a contingency
table depending on the variables selected and the thresholds in case the
variable is numerical. To do so the variables are transformed to factors.
}
\examples{
var1 <- runif(10)
var2 <- runif(10)
df <- data.frame(var1, var2)
df_cont_table(df, 'var1', 0.5, 'var2', c(0.25, 0.5, 0.75))
df_cont_table(df, 'var1', 0.5)

}
