% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_arc}
\alias{draw_arc}
\title{Draw arcs for multiple instances of a subject}
\usage{
draw_arc(x0, y0, x1, y1, p, ggplot_gen = FALSE, cex = 1, col = "black")
}
\arguments{
\item{p}{ggplot object}

\item{ggplot_gen}{logical, if TRUE add the arcs to the ggplot object}

\item{x}{x coordinate}

\item{y}{y coordinate}
}
\value{
Plot the arcs or add it to a ggplot object
}
\description{
Draw arcs for multiple instances of a subject
}
