% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_mz_twin_connection}
\alias{draw_mz_twin_connection}
\title{Draw midpoint MZ twin line}
\usage{
draw_mz_twin_connection(twins, pos, target, i, legh, p, ggplot_gen)
}
\description{
Draw midpoint MZ twin line
}
