% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_polygon}
\alias{draw_polygon}
\title{Draw a polygon for a pedigree}
\usage{
draw_polygon(
  x,
  y,
  p,
  ggplot_gen = FALSE,
  fill = "grey",
  border = NULL,
  density = NULL,
  angle = 45
)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{p}{ggplot object}

\item{ggplot_gen}{logical, if TRUE add the segments to the ggplot object}

\item{fill}{fill color}

\item{border}{border color}

\item{density}{density of shading}

\item{angle}{angle of shading}
}
\value{
Plot the polygon or add it to a ggplot object
}
\description{
Draw a polygon for a pedigree
}
\keyword{internal}
