% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_segment}
\alias{draw_segment}
\title{Draw segments for a Pedigree}
\usage{
draw_segment(
  x0,
  y0,
  x1,
  y1,
  p,
  ggplot_gen,
  col = par("fg"),
  lwd = par("lwd"),
  lty = par("lty")
)
}
\arguments{
\item{x0}{x coordinate of the first point}

\item{y0}{y coordinate of the first point}

\item{x1}{x coordinate of the second point}

\item{y1}{y coordinate of the second point}

\item{p}{ggplot object}

\item{ggplot_gen}{logical, if TRUE add the segments to the ggplot object}

\item{col}{line color}

\item{lwd}{line width}

\item{lty}{line type}
}
\value{
Plot the segments or add it to a ggplot object
}
\description{
Draw segments for a Pedigree
}
\keyword{Pedigree-plot}
\keyword{internal,}
