% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_text}
\alias{draw_text}
\title{Draw text for a Pedigree}
\usage{
draw_text(
  x,
  y,
  label,
  p,
  ggplot_gen = FALSE,
  cex = 1,
  col = NULL,
  adjx = 0,
  adjy = 0
)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{label}{text to be displayed}

\item{p}{ggplot object}

\item{ggplot_gen}{logical, if TRUE add the segments to the ggplot object}

\item{cex}{character expansion of the text}

\item{col}{text color}

\item{adjx}{x adjustment}

\item{adjy}{y adjustment}
}
\value{
Plot the text or add it to a ggplot object
}
\description{
Draw text for a Pedigree
}
\keyword{internal}
