% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{drawbox}
\alias{drawbox}
\title{Draw a box for a pedigree}
\usage{
drawbox(
  x,
  y,
  sex,
  affected,
  status,
  avail,
  mark,
  polylist,
  fill,
  border,
  density,
  angle,
  boxw,
  boxh,
  p_plot,
  id,
  data_tooltips,
  cex,
  ggplot_gen = FALSE,
  tips_names = NULL
)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{sex}{sex of the subject}
}
\value{
Plot the box or add it to a ggplot object
}
\description{
Draw a box for a pedigree
}
