% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{duporder}
\alias{duporder}
\title{Routine to find the duplicate pairs of a subject}
\usage{
duporder(idlist, plist, lev, ped)
}
\arguments{
\item{idlist}{The list of ids}

\item{plist}{The alignment structure}

\item{lev}{The level of the subject}

\item{ped}{The pedigree structure}
}
\value{
A matrix of duplicate pairs
}
\description{
Routine to find the duplicate pairs of a subject
}
\keyword{internal}
