% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unavailable.R
\name{exclude_stray_marryin}
\alias{exclude_stray_marryin}
\title{Exclude stray marry-ins}
\usage{
exclude_stray_marryin(id, dadid, momid)
}
\arguments{
\item{id}{Vector of subject identifiers}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}
}
\value{
Returns a data frame of subject identifiers and their parents.
The data frame is trimmed of any founders who are not parents.
}
\description{
Exclude from a Pedigree any founders who are not parents.
}
\keyword{internal}
