% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unavailable.R
\name{exclude_unavail_founders}
\alias{exclude_unavail_founders}
\title{Exclude unavailable founders}
\usage{
exclude_unavail_founders(id, dadid, momid, avail, missid = NA_character_)
}
\arguments{
\item{id}{Vector of subject identifiers}

\item{missid}{Character defining the missing ids}
}
\value{
Returns a list with the following components:
\itemize{
\item n_trimmed Number of trimmed individuals
\item id_trimmed Vector of IDs of trimmed individuals
\item id Vector of subject identifiers
\item dadid Vector of father identifiers
\item momid Vector of mother identifiers
}
}
\description{
Exclude from a Pedigree any unavailable founders.
}
\keyword{internal}
