% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllAccessors.R
\name{famid<-}
\alias{famid<-}
\title{Famid setter}
\usage{
famid(x) <- value
}
\arguments{
\item{x}{A Ped or Rel object.}

\item{value}{A character or integer vector that should be use
as the new family id}
}
\value{
An updated Ped, Rel object.
}
\description{
Famid setter
}
