% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllAccessors.R
\name{famid}
\alias{famid}
\title{Famid getter}
\usage{
famid(x)
}
\arguments{
\item{x}{A Ped, Rel or Pedigree object.}
}
\value{
A character vector of the family identifiers.
}
\description{
Extract the family identifiers out
of a Ped, Rel or Pedigree object.
If use with a Pedigree object the famid slot of
its own Ped object will be given.
}
