% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_check.R
\name{family_check}
\alias{family_check}
\title{Check family}
\usage{
family_check(obj, ...)
}
\arguments{
\item{famid}{A vector of family identifiers}

\item{id}{A vector of unique subject identifiers}

\item{dadid}{Vector containing the id of the biological father}

\item{momid}{Vector containing the id of the biological mother}

\item{newfam}{The result of a call to \code{make_famid}. If this has already
been computed by the user, adding it as an argument shortens the running
time somewhat.}
}
\value{
a data frame with one row for each unique family id in the
\code{famid} argument. Components of the output are:
\subsection{famid}{

The family id, as entered into the data set
}

\subsection{n}{

Number of subjects in the family
}

\subsection{unrelated}{

Number of them that appear to be unrelated to
anyone else in the entire pedigree set.  This is usually marry-ins with no
children (in the pedigree), and if so are not a problem.
}

\subsection{split}{

Number of unique 'new' family ids.
0 = no one in this 'family' is related to anyone else (not good)
1 = everythings is fine
2+= the family appears to be a set of disjoint trees.
Are you missing some of the people?
}

\subsection{join}{

Number of other families that had a unique
famid, but are actually joined to this one.  0 is the hope.
}
}
\description{
Error check for a family classification
}
\details{
Given a family id vector, also compute the familial grouping from first
principles using the parenting data, and compare the results.

The \code{make_famid} function is used to create a de novo family id from the
parentage data, and this is compared to the family id given in the data.

If there are any joins, then an attribute 'join' is attached.
It will be a matrix with famid as row labels, new-family-id as the columns,
and the number of subjects as entries.
}
\examples{

# use 2 samplepeds
data(sampleped)
pedAll <- with(sampleped, pedigree(id, father, mother, sex,
  affected = cbind(affected, avail), famid = ped
))

## check them giving separate ped ids
fcheck.sep <- with(sampleped, family_check(ped, id, father, mother))
fcheck.sep

## check assigning them same ped id
fcheck.combined <- with(sampleped, family_check(rep(1, nrow(sampleped)),
id, father, mother))
fcheck.combined

# make person 120's father be her son.
sampleped[20, 3] <- 131
fcheck1.bad <- try(
  {
    with(sampleped, family_check(ped, id, father, mother))
  },
  silent = FALSE
)

## fcheck1.bad is a try-error

}
\seealso{
\code{make_famid}, \code{makekinship}
}
\keyword{genetics}
