% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_avail_affected.R
\name{find_avail_affected}
\alias{find_avail_affected}
\title{Find a single person to trim from a pedigree whose is available}
\usage{
find_avail_affected(ped, avail = ped$ped$avail, affstatus = NA)
}
\arguments{
\item{ped}{A pedigree object}

\item{avail}{Vector of availability status (e.g., genotyped) 0/1 or
TRUE/FALSE}

\item{affstatus}{Affection status to search for.}
}
\value{
A list is returned with the following components
\itemize{
\item ped The new pedigree object
\item newAvail Vector of availability status of trimmed individuals
\item idTrimmed Vector of IDs of trimmed individuals
\item isTrimmed logical value indicating whether pedigree has been trimmed
\item bit_size Bit size of the trimmed pedigree
}
}
\description{
Finds one subject from among available non-parents with indicated affection
status.
}
\details{
When used within pedigree.shrink, this function is called with the first
affected indicator, if the affected item in the pedigree is a matrix of
multiple affected indicators.
}
\examples{
data(sampleped)
ped <- pedigree(sampleped)
find_avail_affected(ped, affstatus = 1)
}
\seealso{
\code{\link[=shrink]{shrink()}}
}
