% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_avail_noninform.R
\name{find_avail_noninform}
\alias{find_avail_noninform}
\title{Find uninformative but available subject}
\usage{
find_avail_noninform(ped, avail = ped$ped$avail, missid = "0")
}
\arguments{
\item{ped}{A Pedigree object}

\item{avail}{A numeric vector of availability status of each individual
(e.g., genotyped). The values are:
\itemize{
\item \code{0}  : unavailable
\item \code{1}  : available
\item \code{NA} : availability not known
}}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the Pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{"0"}.}
}
\value{
Vector of subject ids who can be removed by having lowest
informativeness.
}
\description{
Find uninformative but available subject
}
\details{
Find subjects from a Pedigree who are available and uninformative

Identify subjects to remove from a Pedigree who are available but
non-informative.  This is the second step to remove subjects in
pedigree.shrink if the Pedigree does not meet the desired bit size.
}
\examples{
data(sampleped)
ped <- Pedigree(sampleped)
find_avail_noninform(ped)

}
\seealso{
\code{\link[=shrink]{shrink()}}
}
