% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAvailAffected.R
\name{findAvailAffected}
\alias{findAvailAffected}
\title{Find a single person to trim from a pedigree whose is available}
\usage{
findAvailAffected(ped, avail, affstatus)
}
\arguments{
\item{ped}{A pedigree objects, with id (subject ID), findex (father index),
mindex (mother index)}

\item{avail}{Vector of availability status (e.g., genotyped) 0/1 or
TRUE/FALSE}

\item{affstatus}{Vector of affection status 0/1 or TRUE/FALSE.}
}
\value{
A list is returned with the following components
## ped
Dataframe with trimmed subject removed
## idTrimmed
Vector of IDs of trimmed individuals
## isTrimmed
logical value indicating whether pedigree has been trimmed
## bitSize
Bit size of the trimmed pedigree
}
\description{
Finds one subject from among available non-parents with indicated affection
status
}
\details{
When used within pedigree.shrink, this function is called with the first
affected indicator, if the affected item in the pedigree is a matrix of
multiple affected indicators.
}
\seealso{
`pedigree.shrink`
}
