% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAvailNonInform.R
\name{findAvailNonInform}
\alias{findAvailNonInform}
\title{Find uninformative but available subject}
\usage{
findAvailNonInform(ped, avail)
}
\arguments{
\item{ped}{A pedigree object}

\item{avail}{Vector of availability status (e.g. genotyped) 0/1 or
TRUE/FALSE}
}
\value{
Vector of subject ids who can be removed by having lowest
informativeness.
}
\description{
Find uninformative but available subject
}
\details{
Find subjects from a pedigree who are available and uninformative

Identify subjects to remove from a pedigree who are available but
non-informative.  This is the second step to remove subjects in
pedigree.shrink if the pedigree does not meet the desired bit size.
}
\seealso{
`pedigree.shrink`
}
