% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{findsibs}
\alias{findsibs}
\title{Routine to find the siblings of a subject}
\usage{
findsibs(idpos, plist, lev)
}
\arguments{
\item{idpos}{The position of the subject}

\item{plist}{The alignment structure}

\item{lev}{The generation level of the subject}
}
\value{
The positions of the siblings
}
\description{
Routine to find the siblings of a subject
}
\details{
This routine is used by \code{auto_hint()}.
It finds the siblings of a subject.
}
\seealso{
\code{\link[=auto_hint]{auto_hint()}}
}
