% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{findspouse}
\alias{findspouse}
\title{Routine to find the spouse of a subject}
\usage{
findspouse(mypos, plist, lev, ped)
}
\arguments{
\item{mypos}{The position of the subject}

\item{plist}{The alignment structure}

\item{lev}{The level of the subject}

\item{ped}{The pedigree structure}
}
\value{
The position of the spouse
}
\description{
Routine to find the spouse of a subject
}
\keyword{internal}
