% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_parents.R
\name{fix_parents}
\alias{fix_parents}
\alias{fix_parents,character-method}
\alias{fix_parents,character}
\alias{fix_parents,data.frame-method}
\title{Fix details on the parents for children of the Pedigree}
\usage{
fix_parents(obj, ...)

\S4method{fix_parents}{character}(obj, dadid, momid, sex, famid = NULL, missid = NA_character_)

\S4method{fix_parents}{data.frame}(obj, delete = FALSE, filter = NULL, missid = NA_character_)
}
\arguments{
\item{obj}{A data.frame or a vector of the individuals identifiers. If a
dataframe is given it must contain the columns \code{id}, \code{dadid},
\code{momid}, \code{sex} and \code{famid}. famid is optional.}

\item{...}{Additional arguments passed to methods}

\item{sex}{A character, factor or numeric vector corresponding to
the gender of the individuals. This will be transformed to an ordered factor
with the following levels: \code{male} < \code{female} < \code{unknown} < `terminated
The following values are recognized:
\itemize{
\item character() or factor() : "f", "m", "woman", "man", "male", "female",
"unknown", "terminated"
\item numeric() : 1 = "male", 2 = "female", 3 = "unknown", 4 = "terminated"
}}

\item{famid}{Optional famid identification set it to NULL to invalidate.
If used it will modify the ids of the individuals by pasting it with an _.}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the Pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{NA_character_}.}

\item{delete}{Boolean defining if missing parents needs to be:
\itemize{
\item \code{TRUE} : added as a new row
\item \code{FALSE} : be deleted
}}

\item{filter}{Filtering column containing \code{0} or \code{1} for the
rows to kept before proceeding.}
}
\value{
A data.frame with id, dadid, momid, sex as columns with the
relationships fixed.
}
\description{
Fix the sex of parents, add parents that are missing from the Pedigree
Can be used with a dataframe or a vector of the
different individuals informations.
}
\details{
First look to add parents whose ids are given in momid/dadid. Second, fix
sex of parents. Last look to add second parent for children for whom only
one parent id is given.
If a famid vector is given the famid id will be added to the ids of all
individuals (id, dadid, momid) separated by an underscore befor proceeding.
\subsection{Special case for dataframe}{

Check for presence of both parents id in the \code{id} field.
If not both presence behaviour depend of \code{delete} parameter
\itemize{
\item If TRUE then use fix_parents function and merge back the other fields
in the dataframe then set availability to O for non available parents.
\item If FALSE then delete the id of missing parents
}
}
}
\examples{

test1char <- data.frame(
  id = paste('fam', 101:111, sep = ''),
  sex = c('male', 'female')[c(1, 2, 1, 2, 1, 1, 2, 2, 1, 2, 1)],
  father = c(
    0, 0, 'fam101', 'fam101', 'fam101', 0, 0,
    'fam106', 'fam106', 'fam106', 'fam109'
  ),
  mother = c(
    0, 0, 'fam102', 'fam102', 'fam102', 0, 0,
    'fam107', 'fam107', 'fam107', 'fam112'
  )
)
test1newmom <- with(test1char, fix_parents(id, father, mother,
  sex,
  missid = NA_character_
))
newped <- Pedigree(test1newmom)
as.data.frame(newped)

}
\author{
Jason Sinnwell
}
