% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixParents.R
\name{fixParents,data.frame-method}
\alias{fixParents,data.frame-method}
\title{Fix missing parents}
\usage{
\S4method{fixParents}{data.frame}(
  obj,
  delete = FALSE,
  filter = NULL,
  missid = "0",
  id = "id",
  dadid = "dadid",
  momid = "momid",
  sex = "sex",
  family = "family"
)
}
\arguments{
\item{delete}{Boolean defining if missing parents needs to be:
TRUE: added as a new row
FALSE: be deleted}

\item{filter}{Filtering column containing 0 or 1 for the individual to kept}

\item{missid}{The founders are those with no father or mother in the
pedigree.  The \\code{dadid} and \\code{momid} values for these subjects will
either be NA or the value of this variable.  The default for \\code{missid}
is 0 if the \\code{id} variable is numeric, and '' (the empty string)
otherwise.}

\item{id}{Individual id column}

\item{dadid}{Father id column}

\item{momid}{Mother id column}

\item{sex}{Sex column}

\item{family}{Optional family column, set it to NULL to invalidate.}

\item{df}{Dataframe to process}
}
\value{
The same dataframe with the parents ids fixed
}
\description{
Apply fixParents on a dataframe or delete missing parents.
}
\details{
Check for presence of both parents id in the `id` field.
If not both presence behaviour depend of `delete` parameter
If TRUE then use fixParents function and merge back the other fields
in the dataframe then set availability to O for non available parents.
If FALSE then delete the id of missing parents
}
