% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixParents.R
\name{fixParents.character}
\alias{fixParents.character}
\title{Fix missing parents}
\usage{
\method{fixParents}{character}(id, dadid, momid, sex, missid = 0, ...)
}
\description{
Apply fixParents on character vector or delete missing parents
}
