% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixParents.R
\name{fixParents.data.frame}
\alias{fixParents.data.frame}
\title{Fix missing parents}
\usage{
\method{fixParents}{data.frame}(
  df = df,
  delete = FALSE,
  filter = NULL,
  missid = "0",
  id = "id",
  dadid = "dadid",
  momid = "momid",
  sex = "sex",
  ...
)
}
\arguments{
\item{df}{Dataframe to process}

\item{delete}{Boolean defining if missing parents needs to be:
TRUE: added as a new row
FALSE: be deleted}

\item{missid}{The founders are those with no father or mother in the
pedigree.  The \\code{dadid} and \\code{momid} values for these subjects will
either be NA or the value of this variable.  The default for \\code{missid}
is 0 if the \\code{id} variable is numeric, and '' (the empty string)
otherwise.}
}
\value{
The same dataframe with the parents ids fixed
}
\description{
Apply fixParents on a dataframe or delete missing parents.
}
\details{
Check for presence of both parents id in the `id` field.
If not both presence behaviour depend of `delete` parameter
If TRUE then use fixParents function and merge back the other fields
in the dataframe then set availability to O for non available parents.
If FALSE then delete the id of missing parents
}
