% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixParents.R
\name{fixParents.default}
\alias{fixParents.default}
\title{Fix details on the parents for children of the pedigree}
\usage{
\method{fixParents}{default}(x, ...)
}
\description{
Fix the sex of parents, add parents that are missing from the pedigree
}
\details{
First look to add parents whose ids are given in momid/dadid. Second, fix
sex of parents. Last look to add second parent for children for whom only
one parent id is given.
}
\seealso{
\\code{\\link{pedigree}}
}
\author{
Jason Sinnwell
}
