% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixParents.R
\name{fixParents.numeric}
\alias{fixParents.numeric}
\title{Fix missing parents}
\usage{
\method{fixParents}{numeric}(id, dadid, momid, sex, missid = 0, ...)
}
\description{
Apply fixParents on a numeric vector or delete missing parents
}
