% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscore.pedigree.R
\name{fscore.pedigree}
\alias{fscore.pedigree}
\title{Founder score}
\usage{
fscore.pedigree(ped)
}
\arguments{
\item{ped}{}
}
\description{
Attempt at a founder score
}
\seealso{
\code{\link{}}, \code{\link{}}
}
