% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_aff_inds.R
\name{generate_aff_inds}
\alias{generate_aff_inds}
\title{Process the information for affection}
\usage{
generate_aff_inds(
  df,
  col_aff,
  mods_aff = NULL,
  threshold = NULL,
  sup_thres_aff = NULL
)
}
\arguments{
\item{df}{The dataframe to process}

\item{col_aff}{The name of the column containing the values to process as
affection}

\item{mods_aff}{Vector of modality to consider as affected in the case
where the `col_aff` column is a factor.}

\item{threshold}{Numeric value separating the affected and healthy subject
in the case where the `col_aff` column is numeric.}

\item{sup_thres_aff}{Boolean defining if the affected individual are above
the threshold or not. If TRUE, the individuals will be considered affected
if the value of `col_aff` is stricly above the `threshold`. If FALSE, the
individuals will be considered affected if the value of `col_aff` is
stricly under the `threshold`.}
}
\value{
The dataframe with the `affected` column processed accordingly
}
\description{
Process the information for affection
}
\details{
Perform transformation uppon a column given as the one
containing affection status to get an `affected` column usable for
the rest of the script
}
