% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_colors.R
\name{generate_border}
\alias{generate_border}
\title{Process the colors based on affection and availability}
\usage{
generate_border(avail, colors_avail = c("green", "black"))
}
\arguments{
\item{avail}{The vector containing the availability status.
The values need to be numeric and can only be 0, 1 or NA.}

\item{colors_avail}{Set of 2 colors to use for the box's border of an
individual. The first color will be used for available individual (avail
== 1) and the second for the unavailable individual (avail == 0).}
}
\value{
A dataframe containing the scale to use for the availability
status.
}
\description{
Process the colors based on affection and availability
}
\details{
Perform transformation uppon a column given as the one
containing the availability status to compute the border color.
}
