% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_colors.R
\name{generate_colors}
\alias{generate_colors}
\title{Process the colors based on affection and availability}
\usage{
generate_colors(
  df,
  col_aff,
  keep_full_scale = FALSE,
  breaks = 3,
  colors_avail = c("green", "black"),
  colors_aff = c("yellow2", "red"),
  colors_unaff = c("white", "steelblue4")
)
}
\arguments{
\item{df}{The dataframe to process}

\item{col_aff}{The name of the column containing the values to process as
affection}

\item{keep_full_scale}{Boolean defining if the affection values need to
be set as a scale. If `col_aff` is numeric the filling scale will be
calculated based on the value of `col_aff` and the number of breaks given.
If `col_aff` isn't numeric then each levels will get it's own color}

\item{colors_avail}{Set of 2 colors to use for the box's border of an
individual. The first color will be used for available individual (avail
== 1) and the second for the unavailable individual (avail == 0).}

\item{colors_aff}{Set of increasing colors to use for the filling of the
affected individuls.}

\item{colors_unaff}{Set of increasing colors to use for the filling of the
unaffected individuls.}

\item{break}{Number of breaks to use when using full scale with numeric
values. The same number of breaks will be done for values from affected
individuals and unaffected individuals.}
}
\value{
A list of two elements
- The dataframe with the columns `fill` and `border` updated
- List containing the corresponding color scale used for
`col_aff` and availability
}
\description{
Process the colors based on affection and availability
}
\details{
Perform transformation uppon a column given as the one
containing affection status to compute the filling color and use
the `avail` column to compute the border color.
}
