% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_colors.R
\name{generate_fill}
\alias{generate_fill}
\title{Process the colors based on affection}
\usage{
generate_fill(
  values,
  affected,
  labels,
  keep_full_scale = FALSE,
  breaks = 3,
  colors_aff = c("yellow2", "red"),
  colors_unaff = c("white", "steelblue4")
)
}
\arguments{
\item{values}{The vector containing the values to process as affection.}

\item{affected}{The vector containing the affection status.}

\item{labels}{The vector containing the labels to use for the affection.}

\item{keep_full_scale}{Boolean defining if the affection values need to
be set as a scale. If \code{values} is numeric the filling scale will be
calculated based on the values and the number of breaks given.
If \code{values} isn't numeric then each levels will get it's own color}

\item{colors_aff}{Set of increasing colors to use for the filling of the
affected individuls.}

\item{colors_unaff}{Set of increasing colors to use for the filling of the
unaffected individuls.}

\item{break}{Number of breaks to use when using full scale with numeric
values. The same number of breaks will be done for values from affected
individuals and unaffected individuals.}
}
\value{
A list of two elements
\itemize{
\item The processed values column as a numeric factor
\item A dataframe containing the description of each modality of the scale
}
}
\description{
Process the colors based on affection
}
\details{
Perform transformation uppon a column given as the one
containing affection status to compute the filling color.
}
