% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{get_twin_rel}
\alias{get_twin_rel}
\title{Routine to get twin relationships}
\usage{
get_twin_rel(ped)
}
\arguments{
\item{ped}{The pedigree structure}
}
\value{
A list containing components `twinset`, `twinrel` and `twinord`
}
\description{
Routine to get twin relationships
}
\details{
This routine function determine the twin relationships
in a pedigree. It complete the missing twin relationships for
triplets, quads, etc. It also determine the order of the twins
in the pedigree.
It is used by \code{\link{auto_hint}}.
}
