% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructor.R, R/AllGeneric.R
\docType{methods}
\name{Hints}
\alias{Hints}
\alias{Hints,Hints,missing_OR_NULL-method}
\alias{Hints,list,missing_OR_NULL-method}
\alias{Hints,list,missing_OR_NULL}
\alias{Hints,numeric,data.frame-method}
\alias{Hints,numeric,data.frame}
\alias{Hints,numeric,missing_OR_NULL-method}
\alias{Hints,numeric,missing_OR_NULL}
\alias{Hints,missing_OR_NULL,data.frame-method}
\alias{Hints,missing_OR_NULL,data.frame}
\alias{Hints,missing_OR_NULL,missing_OR_NULL-method}
\alias{Hints,missing_OR_NULL,missing_OR_NULL}
\alias{as.list,Hints-method}
\title{Create a Hints object}
\usage{
Hints(horder, spouse)

\S4method{Hints}{Hints,missing_OR_NULL}(horder, spouse)

\S4method{Hints}{list,missing_OR_NULL}(horder, spouse)

\S4method{Hints}{numeric,data.frame}(horder, spouse)

\S4method{Hints}{numeric,missing_OR_NULL}(horder, spouse)

\S4method{Hints}{missing_OR_NULL,data.frame}(horder, spouse)

\S4method{Hints}{missing_OR_NULL,missing_OR_NULL}(horder, spouse)

\S4method{as.list}{Hints}(x)
}
\arguments{
\item{horder}{A named numeric vector with one element per subject in the
Pedigree.  It determines the relative horizontal order of subjects within a
sibship, as well as the relative order of processing for the founder couples.
(For this latter, the female founders are ordered as though
they were sisters).
The names of the vector should be the individual identifiers.}

\item{spouse}{A data.frame with one row per hinted marriage, usually only
a few marriages in a pedigree will need an added hint, for instance reverse
the plot order of a husband/wife pair.
Each row contains the id of the left spouse (i.e. \code{idl}), the id of the
right hand spouse (i.e. \code{idr}), and the anchor (i.e : \code{anchor} :
\code{1} = left, \code{2} = right, \code{0} = either).
Children will preferentially appear under the parents of the anchored spouse.}

\item{from}{A Hints object.}
}
\value{
A Hints object.

A list with the hints informations.
}
\description{
Create a Hints object.
The hints are used to specify the order of the individuals in the pedigree
and to specify the order of the spouses.
You can specify the horder and the spouse in the call to the constructor
independently or together.

Convert a Hints object to a list
}
\examples{
Hints(
  list(
      horder = c("1" = 1, "2" = 2, "3" = 3),
      spouse = data.frame(
          idl = c("1", "2"),
          idr = c("2", "3"),
          anchor = c(1, 2)
  )
)
Hints(
  horder = c("1" = 1, "2" = 2, "3" = 3),
  spouse = data.frame(
      idl = c("1", "2"),
      idr = c("2", "3"),
      anchor = c(1, 2)
  )
)
Hints(
  horder = c("1" = 1, "2" = 2, "3" = 3)
)
Hints(
    spouse = data.frame(
        idl = c("1", "2"),
        idr = c("2", "3"),
        anchor = c(1, 2)
    )
)
Hints()
}
\seealso{
\code{\link[=Pedigree]{Pedigree()}}
}
