% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_founder}
\alias{is_founder}
\title{Check wich individuals are founders}
\usage{
is_founder(momid, dadid, missid = NA_character_)
}
\value{
A vector of boolean of the same size as \code{dadid} and \code{momid}
with TRUE if the individual has no parents (i.e is a founder) and FALSE
otherwise.
}
\description{
Check which individuals are founders.
}
\keyword{internal}
