% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_parent}
\alias{is_parent}
\title{Check wich individuals are parents}
\usage{
is_parent(obj, ...)
}
\arguments{
\item{obj}{A character vector with the id of the individuals or a
\code{data.frame} with all the informations in corresponding columns.}

\item{id}{A vector of each subjects identifiers}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the Pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{NA_character_}.}
}
\value{
A vector of boolean of the same size as \code{id}
with TRUE if the individual is a parent and FALSE otherwise
}
\description{
Check which individuals are parents.
}
\keyword{internal}
