% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeValidity.R
\name{is_valid}
\alias{is_valid}
\title{Pedigree validity method.}
\usage{
is_valid(object)
}
\arguments{
\item{object}{A Pedigree object.}
}
\value{
A logical value or a character vector with the errors.
}
\description{
Check if the Pedigree object is valid.
}
\details{
It will check :
the fields of the slots
the values in the columns of the ped, rel and scale slot
}
\keyword{internal}
