% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllValidity.R
\name{is_valid_hints}
\alias{is_valid_hints}
\title{Check if the Hints are valid}
\usage{
is_valid_hints(object)
}
\arguments{
\item{object}{A Hints object.}
}
\value{
A character vector with the errors or \code{TRUE} if no errors.
}
\description{
Check if horder and spouse slots are valid (i.e. horder is numeric and
spouse is a matrix with 3 columns).
Check if the spouse matrix is valid (i.e. only numeric values).
}
\keyword{internal}
