% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllValidity.R
\name{is_valid_ped}
\alias{is_valid_ped}
\title{Check if a Ped object is valid}
\usage{
is_valid_ped(object)
}
\arguments{
\item{object}{A Ped object.}
}
\value{
A character vector with the errors or \code{TRUE} if no errors.
}
\description{
Multiple checks are done here
}
\details{
\enumerate{
\item Check that the ped ids slots have the right values
\item Check that the sex, steril, status, avail and affected slots have the
right values
\item Check that dad are male and mom are female
\item Check that individuals have both parents or none
}
}
\keyword{internal}
