% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllValidity.R
\name{is_valid_pedigree}
\alias{is_valid_pedigree}
\title{Check if a Pedigree object is valid}
\usage{
is_valid_pedigree(object)
}
\arguments{
\item{object}{A Ped object.}
}
\value{
A character vector with the errors or \code{TRUE} if no errors.
}
\description{
Multiple checks are done here
}
\details{
\enumerate{
\item Check that the all Rel id are in the Ped object
\item Check that twins have same parents
\item Check that MZ twins have same sex
\item Check that all columns used in scales are in the Ped object
\item Check that all fill & border modalities are in the Ped object column
\item Check that all id used in Hints object are in the Ped object
\item Check that all spouse in Hints object are male / female
}
}
\keyword{internal}
