% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllValidity.R
\name{is_valid_rel}
\alias{is_valid_rel}
\title{Check if a Rel object is valid}
\usage{
is_valid_rel(object)
}
\arguments{
\item{object}{A Ped object.}
}
\value{
A character vector with the errors or \code{TRUE} if no errors.
}
\description{
Multiple checks are done here
}
\details{
\enumerate{
\item Check that the "id1", "id2", "code", "famid" slots exist
\item Check that the "code" slots have the right values
(i.e. "MZ twin", "DZ twin", "UZ twin", "Spouse")
\item Check that all "id1" are different to "id2"
\item Check that all "id1" are smaller than "id2"
\item Check that no duplicate relation are present
}
}
\keyword{internal}
