% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllValidity.R
\name{is_valid_scales}
\alias{is_valid_scales}
\title{Check if the Scales are valid}
\usage{
is_valid_scales(object)
}
\arguments{
\item{object}{A Scales object.}
}
\value{
A character vector with the errors or \code{TRUE} if no errors.
}
\description{
Check if the fill and border slots are valid (i.e. they have the right
columns).
}
\keyword{internal}
