% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship2-package.R
\docType{package}
\name{kinship2-package}
\alias{kinship2}
\alias{kinship2-package}
\title{kinship2: Pedigree Functions}
\description{
Routines to handle family data with a pedigree object. The initial purpose was to create correlation structures that describe family relationships such as kinship and identity-by-descent, which can be used to model family data in mixed effects models, such as in the coxme function. Also includes a tool for pedigree drawing which is focused on producing compact layouts without intervention. Recent additions include utilities to trim the pedigree object with various criteria, and kinship for the X chromosome.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cran.r-project.org/package=kinship2}
  \item Report bugs at \url{https://github.com/LouisLeLezet/kinship2/issues}
}

}
\author{
\strong{Maintainer}: Louis Le Nezet \email{louislenezet@gmail.com} (\href{https://orcid.org/0009-0000-0202-2703}{ORCID})

Authors:
\itemize{
  \item Jason Sinnwell \email{sinnwell.jason@mayo.edu}
  \item Terry Therneau
}

Other contributors:
\itemize{
  \item Daniel Schaid [contributor]
  \item Elizabeth Atkinson [contributor]
  \item Carly Mester [contributor]
}

}
\keyword{internal}
