% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship2-package.R
\docType{package}
\name{kinship2-package}
\alias{kinship2}
\alias{kinship2-package}
\title{The kinship2 package for pedigree data}
\description{
The kinship2 package for pedigree data
}
\details{
The package download, NEWS, and README are available on CRAN:
\\url{https://cran.r-project.org/package=kinship2}
}
\section{Functions}{

Below are listed some of the most widely used functions available
in arsenal:

\code{\link[=pedigree]{pedigree()}}: Contstructor of the pedigree class,
given identifiers, sex, affection status(es), and special relationships

\code{\link[=kinship]{kinship()}}: Calculates the kinship matrix, the probability
having an allele sampled from two individuals be the same via IBD.

\code{\link[=ped_to_plotdf]{ped_to_plotdf()}} and \code{\link[=plot_fromdf]{plot_fromdf()}}: Plot method for a pedigree object.
Allows extra information to be included in the id under the plot symbol

\code{\link[=shrink]{shrink()}}: Shrink a pedigree to a specific bit size,
removing non-informative members first.

\code{\link[=bit_size]{bit_size()}}: Approximate the output from SAS's \verb{PROC FREQ}
procedure when using the \verb{/list} option of the \code{TABLE} statement.
}

\section{Data}{

\code{\link[=sampleped]{sampleped()}}: Pedigree example data sets with two pedigrees
\code{\link[=minnbreast]{minnbreast()}}: Larger cohort of pedigrees from MN breast cancer
study
}

\examples{
library(kinship2)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://cran.r-project.org/package=kinship2}
  \item Report bugs at \url{https://github.com/LouisLeLezet/kinship2/issues}
}

}
\author{
\strong{Maintainer}: Louis Le Nezet \email{louislenezet@gmail.com} (\href{https://orcid.org/0009-0000-0202-2703}{ORCID})

Authors:
\itemize{
  \item Jason Sinnwell \email{sinnwell.jason@mayo.edu}
  \item Terry Therneau
}

Other contributors:
\itemize{
  \item Daniel Schaid [contributor]
  \item Elizabeth Atkinson [contributor]
  \item Carly Mester [contributor]
}

}
\keyword{internal}
