% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{length,Ped-method}
\alias{length,Ped-method}
\title{Compute the length of a Ped object}
\usage{
\S4method{length}{Ped}(x)
}
\arguments{
\item{x}{A Ped object.}
}
\value{
The number of individuals in the Ped object.
}
\description{
Compute the length of a Ped object
}
