% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{length,Pedigree-method}
\alias{length,Pedigree-method}
\title{Compute the length of a Pedigree object}
\usage{
\S4method{length}{Pedigree}(x)
}
\arguments{
\item{x}{A Pedigree object.}
}
\value{
The number of individuals in the Pedigree object.
}
\description{
Compute the length of a Pedigree object
}
