% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_famid.R
\docType{methods}
\name{make_famid}
\alias{make_famid}
\alias{make_famid,character-method}
\alias{make_famid,character}
\alias{make_famid,Pedigree-method}
\alias{make_famid,Pedigree}
\title{Get family id}
\usage{
make_famid(obj, ...)

\S4method{make_famid}{character}(obj, dadid, momid)

\S4method{make_famid}{Pedigree}(obj)
}
\arguments{
\item{obj}{A Pedigree object or a vector of subject identifiers.}

\item{...}{Additional arguments passed to methods}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}
}
\value{
\subsection{When used with a character vector}{

An integer vector giving family groupings
}

\subsection{When used with a Pedigree object}{

An updated Pedigree object with the family id added
}
}
\description{
Construct a family id from Pedigree information
}
\details{
Create a vector of length n, giving the family 'tree' number of each
subject.  If the Pedigree is totally connected, then everyone will end up in
tree 1, otherwise the tree numbers represent the disconnected subfamilies.
Singleton subjects give a zero for family number.
}
\examples{
data(sampleped)
ped1 <- Pedigree(sampleped[,-1])
make_famid(ped1)
}
\seealso{
\code{\link[=kinship]{kinship()}}
}
