% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makefamid.R
\name{makefamid,Pedigree-method}
\alias{makefamid,Pedigree-method}
\title{Get family id}
\usage{
\S4method{makefamid}{Pedigree}(obj)
}
\arguments{
\item{ped}{a pedigree object}
}
\value{
An updated pedigree object with the family id added
}
\description{
Construct a family id from pedigree information
}
\details{
Create a vector of length n, giving the family 'tree' number of each
subject.  If the pedigree is totally connected, then everyone will end up in
tree 1, otherwise the tree numbers represent the disconnected subfamilies.
Singleton subjects give a zero for family number.

This command is depricated. The kinship command now can be applied directly
to pedigreeList objects.
}
\seealso{
`kinship`
}
\author{
Terry Therneau
}
\keyword{genetics}
