% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makekinship.R
\name{makekinship}
\alias{makekinship}
\title{Create a sparse kinship matrix}
\usage{
makekinship(famid, id, dadid, momid, unrelated = 0)
}
\arguments{
\item{famid}{a vector of family identifiers}

\item{id}{a vector of unique subject identifiers}

\item{dadid}{for each subject, the identifier of their biolgical father}

\item{momid}{for each subject, the identifier of thier biological mother}

\item{unrelated}{subjects with this family id are considered to be unrelated
singletons, i.e., not related to each other or to anyone else.}
}
\value{
a sparse kinship matrix of class `bdsmatrix`
}
\description{
Compute the overall kinship matrix for a collection of families, and store
it efficiently.
}
\details{
This command is deprecated.  The kinship command now can be applied directly
to pedigreeList objects.
}
\examples{

# Data set from a large family study of breast cancer
#  there are 26050 subjects in the file, from 426 families

data(minnbreast)
table(minnbreast$sex)
# F     M
# 12818 13502

length(unique(minnbreast$famid))
# 426

kin1 <- with(minnbreast, makekinship(famid, id, fatherid, motherid))
dim(kin1)
# 28081 28081

class(kin1)
# 'dsCMatrix'

# kinship matrix for the females only
femid <- minnbreast$id[minnbreast$sex == 'FALSE']
femindex <- !is.na(match(dimnames(kin1)[[1]], femid))
kin2 <- kin1[femindex, femindex]

# Note that 'femindex <- match(femid, dimnames(kin1)[[1]])' is wrong, since
#  then kin1[femindex, femindex] might improperly reorder the rows/cols
#  (if families were not contiguous in minnbreast).
# However sort(match(femid, dimnames(kin1)[[1]])) would be okay.

}
\seealso{
`kinship`, `makefamid`
}
\keyword{genetics}
