% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_kin_inf.R
\docType{methods}
\name{max_kin_inf}
\alias{max_kin_inf}
\alias{max_kin_inf,character-method}
\alias{max_kin_inf,character}
\alias{max_kin_inf,data.frame-method}
\alias{max_kin_inf,data.frame}
\alias{max_kin_inf,Pedigree-method}
\alias{max_kin_inf,Pedigree}
\title{Kinship computation}
\usage{
max_kin_inf(obj, ...)

\S4method{max_kin_inf}{character}(obj, dadid, momid, sex, avail, affected, informative = "AvAf")

\S4method{max_kin_inf}{data.frame}(obj, informative = "AvAf")

\S4method{max_kin_inf}{Pedigree}(obj, informative = "AvAf", reset = FALSE, ...)
}
\arguments{
\item{obj}{A pedigree object, a dataframe or a vector of the individuals}

\item{...}{Other parameters passed to \code{\link[=is_informative]{is_informative()}}
when using a pedigree object}

\item{dadid}{A vector of individuals dadid}

\item{momid}{A vector of individuals momid}

\item{sex}{A vector of the individuals sex}

\item{avail}{A vector of individuals availability (0, 1 or NA)}

\item{affected}{A vector of individuals affected status (0, 1 or NA)}

\item{informative}{Informative individuals selection can take 3 values:
'AvAf' (available and affected),
'AvOrAf' (available or affected),
'Av' (available only),
'Af' (affected only),
'All' (all individuals)
or a numeric vector of individuals id
or a boolean}

\item{reset}{If TRUE, the kinship column is reset}
}
\value{
\subsection{When obj is a vector or a dataframe}{

A vector of kinship degree
}

\subsection{When obj is a Pedigree}{

The Pedigree object with a new column named 'kin' containing the kinship
degree.
}
}
\description{
Compute the kinship between the informative individuals and
all the others.
}
\examples{
data(sampleped)
ped <- pedigree(sampleped)
max_kin_inf(ped, column = "affected_aff")$ped

}
