% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_inf.R
\docType{methods}
\name{min_dist_inf}
\alias{min_dist_inf}
\alias{min_dist_inf,character-method}
\alias{min_dist_inf,character}
\alias{min_dist_inf,Pedigree-method}
\alias{min_dist_inf,Pedigree}
\alias{min_dist_inf,Ped-method}
\alias{min_dist_inf,Ped}
\title{Minimum distance to the informative individuals}
\usage{
min_dist_inf(obj, ...)

\S4method{min_dist_inf}{character}(obj, dadid, momid, sex, avail, affected, informative = "AvAf")

\S4method{min_dist_inf}{Pedigree}(
  obj,
  col_aff = NULL,
  informative = "AvAf",
  missid = NA_character_,
  reset = FALSE,
  ...
)

\S4method{min_dist_inf}{Ped}(
  obj,
  col_aff = NULL,
  informative = "AvAf",
  missid = NA_character_,
  reset = FALSE,
  ...
)
}
\arguments{
\item{obj}{A pedigree object or a vector of subject identifiers.}

\item{...}{Additional arguments passed to methods}

\item{sex}{A character, factor or numeric vector corresponding to
the gender of the individuals. This will be transformed to an ordered factor
with the following levels: \code{male} < \code{female} < \code{unknown} < `terminated
The following values are recognized:
\itemize{
\item character() or factor() : "f", "m", "woman", "man", "male", "female",
"unknown", "terminated"
\item numeric() : 1 = "male", 2 = "female", 3 = "unknown", 4 = "terminated"
}}

\item{avail}{A logical vector with the availability status of the
individuals
(i.e. \code{FALSE} = not available, \code{TRUE} = available, \code{NA} = unknown).}

\item{affected}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = unaffected, \code{TRUE} = affected, \code{NA} = unknown).}

\item{informative}{Informative individuals selection can take 5 values:
\itemize{
\item 'AvAf' (available and affected),
\item 'AvOrAf' (available or affected),
\item 'Av' (available only),
\item 'Af' (affected only),
\item 'All' (all individuals)
\item A numeric/character vector of individuals id
\item A boolean
}}

\item{col_aff}{A character vector with the name of the column to be used
for the affection status.
If \code{affected} is a data.frame, it will be overwritten by the column
names of the data.frame.}

\item{missid}{A character vector with the missing values identifiers.
All the id, dadid and momid corresponding to those values will be set
to \code{NA_character_}.}

\item{reset}{If TRUE, the \code{kin} and if \code{isinf} columns is reset}
}
\value{
\subsection{When obj is a vector}{

A vector of the minimum distance between the informative individuals
and all the others corresponding to the order of the individuals in the
\code{obj} vector.
}

\subsection{When obj is a Pedigree}{

The Pedigree object with a new column named 'kin' containing the kinship
degree.
}
}
\description{
Compute the minimum distance between the informative
individuals and all the others.
This distance is a transformation of the maximum kinship degree between
the informative individuals and all the others.
This transformation is done by taking the log2 of the inverse of the
maximum kinship degree.

\eqn{minDist = log2(1 / \max(kinship))}

Therefore, the minimum distance is 0 when the maximum kinship is 1 and
is infinite when the maximum kinship is 0. For siblings, the kinship value
is 0.5 and the minimum distance is 1. Each time the kinship degree is divided
by 2, the minimum distance is increased by 1.
}
\examples{
data(sampleped)
ped <- Pedigree(sampleped)
min_dist_inf(ped, col_aff = "affection_mods")$ped

}
\seealso{
\code{\link[=kinship]{kinship()}}
}
