% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{methods}
\name{myfunction}
\alias{myfunction}
\alias{myfunction,character-method}
\alias{myfunction,character}
\alias{myfunction,Pedigree-method}
\alias{myfunction,Pedigree}
\title{My function}
\usage{
myfunction(obj, ...)

## S4 method for signature 'character'

myfunction(dadid, momid, missid = "0")

## S4 method for signature 'Pedigree'

myfunction(ped, missid = "0")
}
\arguments{
\item{obj}{An object either a character vector or a pedigree}

\item{...}{Arguments to be passed to methods}

\item{momid}{A character vector}

\item{missid}{Character defining the missing ids}

\item{dadid}{A character vector}

\item{ped}{A pedigree object}
}
\value{
A character vector with the parents ids

A pedigree with the parents ids
}
\description{
My function
}
