% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructor.R
\name{na_to_length}
\alias{na_to_length}
\title{NA to specific length}
\usage{
na_to_length(x, temp, value)
}
\arguments{
\item{x}{The vector to check.}

\item{temp}{A template vector to use to determine the length.}

\item{value}{The value to use to fill the vector.}
}
\value{
A vector with the same length as temp.
}
\description{
NA to specific length
}
\keyword{internal}
