% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\name{norm_data}
\alias{norm_data}
\title{Normalise dataframe}
\usage{
norm_data(df, na_strings = "NA")
}
\arguments{
\item{df}{The dataframe to process}

\item{na_strings}{Vector of strings to be considered as NA values}
}
\value{
List of two dataframe
`norm` is the dataframe with all the individuals without any errors.
`errors` is the dataframe containing the individuals with errors.
}
\description{
Normalise dataframe for pedigree object
}
\details{
Normalise a dataframe and check for columns correspondance
to be able to use it as an input to create pedigree object.
Multiple test are done and errors are checked.
Sex is calculated based in the `Gender` column the following notations
are accepted: f, woman, female, 2 and m, man, male, 1.
The `Sterilisation` column need to be a boolean either TRUE, FALSE or "NA".
Will be considered available any individual with no "NA" values in the
`Availability` column.
Duplicated `IndID` will nullify the relationship of the individual.
All individuals with errors will be remove from the dataframe and will
be transfered to the error dataframe.
}
\examples{
data(ped_data)
norm_data(ped_data)

}
