% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\name{norm_rel}
\alias{norm_rel}
\title{Normalise relationship dataframe}
\usage{
norm_rel(rel_df, na_strings = c("NA", ""), missid = "0")
}
\arguments{
\item{rel_df}{The dataframe to process}

\item{na_strings}{Vector of strings to be considered as NA values}

\item{missid}{The missing id value}
}
\value{
A dataframe with the errors identified
}
\description{
Normalise relationship dataframe for pedigree object
}
