% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_child.R
\name{num_child}
\alias{num_child}
\title{Number of child}
\usage{
num_child(df, relation = NULL)
}
\arguments{
\item{df}{Dataframe processed for pedigree computation with the columns
`id`, `dadid` and `momid` available}

\item{relation}{A matrix with 3 required columns (id1, id2, code) specifying
special relationship between pairs of individuals.
Codes: 1=Monozygotic twin, 2=Dizygotic twin, 3=Twin of unknown zygosity,
4=Spouse.}
}
\value{
A dataframe with the columns `num_child_dir`, `num_child_ind` and
`num_child_tot` giving respectively the direct, indirect and total number
of child.
}
\description{
Compute the number of child per individual
}
\details{
Compute the number of direct child but also the number
of indirect child given by the ones related with the linked spouses.
If a relation ship matrix is given, then even if no children is present
between 2 spouses, the indirect childs will still be added.
}
