% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{paste0max}
\alias{paste0max}
\title{Print0 to max}
\usage{
paste0max(x, max = 5, ...)
}
\description{
Print0 the elements inside a vector until a maximum is reached.
}
\details{
$param x A vector.
$param max The maximum number of elements to print.
$param ... Additional arguments passed to print0

$export
}
