% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Ped-class}
\alias{Ped-class}
\title{S4 class to represent the identity informations in a Pedigree.}
\value{
A Ped object.
}
\description{
A Ped object is a list of identity informations
of the individuals in the pedigree.
It is used to create a Pedigree object.
The minimal needed informations are \code{id}, \code{dadid}, \code{momid} and \code{sex}.
If a \code{family} is provided, the individuals \code{id} will be aggregated
to the \code{family} character to ensure the uniqueness of the \code{id}.
The other slots are used to store recognized informations.
Additional columns can be added to the Ped object and will be
stored in the \code{meta} slot of the Ped object.
}
\section{Slots}{

\describe{
\item{\code{id}}{A character vector with the id of the individuals.}

\item{\code{dadid}}{A character vector with the id of the father of the individuals.}

\item{\code{momid}}{A character vector with the id of the mother of the individuals.}

\item{\code{sex}}{A factor vector for the sex of the individuals (i.e. \code{male},
\code{female}, \code{unknown} or \code{terminated}).}

\item{\code{family}}{A character vector with the family of the individuals.}

\item{\code{steril}}{A numeric vector with the sterilisation status of the
individuals (i.e. \code{0} = not sterilised, \code{1} = sterilised, \code{NA} = unknown).}

\item{\code{status}}{A numeric vector with the affection status of the
individuals (i.e. \code{0} = alive, \code{1} = dead, \code{NA} = unknown).}

\item{\code{avail}}{A numeric vector with the availability status of the
individuals (i.e. \code{0} = not available, \code{1} = available, \code{NA} = unknown).}

\item{\code{affected}}{A numeric vector with the affection status of the
individuals (i.e. \code{0} = not affected, \code{1} = affected, \code{NA} = unknown).}

\item{\code{useful}}{A numeric vector with the usefulness status of the
individuals (i.e. \code{0} = not useful, \code{1} = useful).}

\item{\code{kin}}{A numeric vector with minimal kinship value between the
individuals and the useful individuals.}

\item{\code{num_child_total}}{A numeric vector with the total number of children
of the individuals.}

\item{\code{num_child_direct}}{A numeric vector with the number of children
of the individuals.}

\item{\code{num_child_indirect}}{A numeric vector with the number of children
of the individuals.}
}}

\seealso{
\code{\link[=Pedigree]{Pedigree()}}
}
