% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped2df.R
\name{ped2df}
\alias{ped2df}
\title{Convert ped to data.frame}
\usage{
ped2df(ped)
}
\arguments{
\item{ped}{}
}
\value{
A dataframe containing the following columns:affected
- id
- findex
- mindex
- sex
- affected (optional)
- status (optional)
}
\description{
Convert a pedigree object to adata.frame
}
\examples{
data(sample.ped)
ped <- with(sample.ped,pedigree(id, father, mother, sex))
ped2df(ped)
}
\seealso{
\code{\link{}}, \code{\link{}}
}
