% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_to_legdf.R
\docType{methods}
\name{ped_to_legdf}
\alias{ped_to_legdf}
\title{Convert a Pedigree to a legend data frame of element to plot}
\usage{
ped_to_legdf(ped, boxh = 1, boxw = 1, cex = 1, adjx = 0, adjy = 0)
}
\arguments{
\item{ped}{A Pedigree object}

\item{boxh}{Height of the legend boxes}

\item{boxw}{Width of the legend boxes}

\item{cex}{character expansion of the text}

\item{adjx}{default=0.  Controls the horizontal text adjustment of
the labels in the legend.}

\item{adjy}{default=0.  Controls the vertical text adjustment
of the labels in the legend.}
}
\value{
A list containing the legend data frame and the user coordinates.
}
\description{
Convert a Pedigree to a legend data frame for it to
be plotted with afterwards with \code{\link[=plot_fromdf]{plot_fromdf()}}.
}
\examples{
data("sampleped")
ped <- Pedigree(sampleped)
leg_df <- ped_to_legdf(ped)
summary(leg_df$leg_df)
plot_fromdf(leg_df$leg_df)
}
