% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_to_plotdf.R
\name{ped_to_plotdf}
\alias{ped_to_plotdf}
\title{Convert a pedigree to a data frame of element to plot}
\usage{
ped_to_plotdf(
  ped,
  packed = FALSE,
  width = 10,
  align = c(1.5, 2),
  subreg = NULL,
  cex = 0.5,
  symbolsize = cex,
  pconnect = 0.5,
  branch = 0.6,
  mark = TRUE,
  label = NULL,
  ...
)
}
\arguments{
\item{ped}{A pedigree object}

\item{packed}{Should the pedigree be compressed, i.e., allow diagonal
lines connecting parents to children in order to have a smaller overall
width for the plot.}

\item{width}{for a packed output, the minimum width of the plot, in
inches.}

\item{align}{for a packed pedigree, align children under parents \code{TRUE}, to
the extent possible given the page width, or align to to the left margin
\code{FALSE}. This argument can be a two element vector, giving the alignment
parameters, or a logical value.  If \code{TRUE}, the default is \code{c(1.5, 2)}, or
numeric the routine \code{alignped4()} will be called.}

\item{subreg}{4-element vector for (min x, max x, min depth, max depth),
used to edit away portions of the plot coordinates returned by
align.pedigree}

\item{cex}{Size of the text}

\item{symbolsize}{Size of the symbols}

\item{pconnect}{when connecting parent to children the program will try to
make the connecting line as close to vertical as possible, subject to it
lying inside the endpoints of the line that connects the children by at
least \code{pconnect} people.  Setting this option to a large number will
force the line to connect at the midpoint of the children.}

\item{branch}{defines how much angle is used to connect various levels of
nuclear families.}

\item{mark}{if TRUE, add a mark to each box corresponding to the value of
the affection column for each filling scale.}

\item{label}{if not NULL, add a label to each box corresponding to the
value of the column given.}

\item{...}{Other arguments passed to \code{\link[=set_plot_area]{set_plot_area()}}.}
}
\value{
A list containing the data frame and the user coordinates.
}
\description{
Convert a pedigree to a data frame of element to plot
}
