% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/AllAccessors.R
\docType{class}
\name{Pedigree-class}
\alias{Pedigree-class}
\alias{famid,Pedigree-method}
\title{S4 class to represent a pedigree.}
\value{
A Pedigree object.
}
\description{
A pedigree is a ensemble of individuals linked to each other into
a family tree.
}
\details{
They are created from a data.frame containing the individuals informations
and a relation ship data.frame for the special links between individuals.
A list of scales can be provided to create a legend.
To create a Pedigree object, use the function
\code{\link[=Pedigree]{Pedigree()}}.
}
\section{Slots}{

\describe{
\item{\code{ped}}{A Ped object for the identity informations. See \code{\link[=Ped]{Ped()}} for
more informations.}

\item{\code{rel}}{A Rel object for the special relationships. See \code{\link[=Rel]{Rel()}} for
more informations.}

\item{\code{scales}}{A Scales object for the filling and bordering
colors used in the plot. See \code{\link[=Scales]{Scales()}} for more informations.}

\item{\code{hints}}{A Hints object for the ordering of the
individuals in the plot. See \code{\link[=Hints]{Hints()}} for more informations.}
}}

\section{Accessors}{

\itemize{
\item \code{famid(x)} : Get the family identifiers of a Pedigree object. This
function is a wrapper around \code{famid(ped(x))}.
}
}

\seealso{
\code{\link[=Pedigree]{Pedigree()}}, \code{\link[=Ped]{Ped()}}, \code{\link[=Rel]{Rel()}}, \code{\link[=Scales]{Scales()}}, \code{\link[=Hints]{Hints()}}
}
