% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pedigree-package.R
\docType{package}
\name{Pedigree-package}
\alias{Pedigree}
\alias{Pedigree-package}
\title{The Pedigree package for pedigree data}
\description{
The Pedigree package for pedigree data an updated package of the
\code{kinship2} package.
The \code{kinship2} package was originally
written by Terry Therneau and Jason Sinnwell.
The \code{Pedigree} package is a fork of the
\code{kinship2} package with
additional functionality and bug fixes.
}
\details{
The package download, NEWS, and README are available on CRAN:
\\url{https://cran.r-project.org/package=kinship2}
}
\section{Functions}{

Below are listed some of the most widely used functions available
in arsenal:

\code{\link[=pedigree]{pedigree()}}: Contstructor of the pedigree class,
given identifiers, sex, affection status(es), and special relationships

\code{\link[=kinship]{kinship()}}: Calculates the kinship matrix, the
probability having an allele sampled from two individuals
be the same via IBD.

\code{\link[=ped_to_plotdf]{ped_to_plotdf()}} : Method to transform a pedigree
object into a dataframe of graphical elements.
Allows extra information to be included in the id under the plot symbol

\code{\link[=plot_fromdf]{plot_fromdf()}} : Method to plot a pedigree from a
dataframe of graphical elements.

\code{\link[=shrink]{shrink()}}: Shrink a pedigree to a specific bit size,
removing non-informative members first.

\code{\link[=bit_size]{bit_size()}}: Approximate the output from SAS's
\verb{PROC FREQ} procedure when using the \verb{/list}
option of the \code{TABLE} statement.
}

\section{Data}{

\itemize{
\item \code{\link[=sampleped]{sampleped()}}: Pedigree example data sets
with two pedigrees
\item \code{\link[=minnbreast]{minnbreast()}}: Larger cohort of pedigrees
from MN breast cancer study
}
}

\examples{
library(Pedigree)

}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/LouisLeLezet/Pedigree/issues}
}

}
\author{
\strong{Maintainer}: Louis Le Nézet \email{louislenezet@gmail.com} (\href{https://orcid.org/0009-0000-0202-2703}{ORCID})

Authors:
\itemize{
  \item Jason Sinnwell \email{sinnwell.jason@mayo.edu}
  \item Terry Therneau
}

Other contributors:
\itemize{
  \item Daniel Schaid [contributor]
  \item Elizabeth Atkinson [contributor]
  \item Louis Le Nezet [contributor]
}

}
\keyword{internal}
