% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_hint.R
\name{permute}
\alias{permute}
\title{Generate all possible permutation}
\usage{
permute(x)
}
\arguments{
\item{x}{A vector of length \strong{n}}
}
\value{
A matrix with \strong{n} cols and \strong{n!} rows
}
\description{
Given a vector of length \strong{n}, generate all possible permutations of
the numbers 1 to \strong{n}.
This is a recursive routine, and is not very efficient.
}
\keyword{auto_hint}
\keyword{internal,}
