% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{polyfun}
\alias{polyfun}
\title{Generate a polygonal element}
\usage{
polyfun(nslice, coor)
}
\arguments{
\item{nslice}{number of slices in the polygon}

\item{coor}{Element form which to generate the polygon
containing x and y coordinates and theta}
}
\value{
a list of x and y coordinates
}
\description{
Generate a polygonal element
}
