% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{polygons}
\alias{polygons}
\title{Create a list of the different polygonal elements}
\usage{
polygons(naffection = 1)
}
\arguments{
\item{naffection}{number of affection status}
}
\value{
a list of polygonal elements with x, y coordinates
and theta
}
\description{
Create a list of the different polygonal elements
}
