% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Rel-class}
\alias{Rel-class}
\title{S4 class to represent the special relationships in a Pedigree.}
\value{
A Rel object.
}
\description{
A Rel object is a list of special relationships
between individuals in the pedigree.
It is used to create a Pedigree object.
The minimal needed informations are \code{id1}, \code{id2} and \code{code}.
If a \code{family} is provided, the individuals \code{id} will be aggregated
to the \code{family} character to ensure the uniqueness of the \code{id}.
}
\section{Slots}{

\describe{
\item{\code{id1}}{A character vector with the id of the first individual.}

\item{\code{id2}}{A character vector with the id of the second individual.}

\item{\code{code}}{A character vector with the code of the special relationship.
(i.e. \verb{MZ twin}, \verb{DZ twin}, \verb{UZ twin} or \code{Spouse}).}

\item{\code{family}}{A character vector with the family of the individuals.}
}}

\seealso{
\code{\link[=Pedigree]{Pedigree()}}
}
