% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/AllConstructor.R, R/AllGeneric.R
\docType{class}
\name{Rel-class}
\alias{Rel-class}
\alias{Rel}
\alias{Rel,data.frame-method}
\alias{Rel,data.frame}
\alias{Rel,character-method}
\alias{Rel,character}
\alias{summary,Rel-method}
\alias{show,Rel-method}
\alias{as.list,Rel-method}
\alias{as.data.frame,Rel-method}
\title{S4 class to represent the special relationships in a Pedigree.}
\usage{
Rel(obj, ...)

\S4method{Rel}{data.frame}(obj)

\S4method{Rel}{character}(obj, id2, code, family = NA)

\S4method{summary}{Rel}(object)

\S4method{show}{Rel}(object)

\S4method{as.list}{Rel}(x)

\S4method{as.data.frame}{Rel}(x)
}
\arguments{
\item{obj}{A character vector with the id of the first individuals of each
pairs or a \code{data.frame} with all the informations in corresponding columns.}

\item{id2}{A character vector with the id of the second individuals of each
pairs}

\item{code}{An ordered factor vector with the code of the special
relationship (i.e. \verb{MZ twin} < \verb{DZ twin} < \verb{UZ twin} < \code{Spouse}).}

\item{family}{A character vector with the family of the individuals.}

\item{object}{A Rel object.}

\item{from}{A Rel object.}
}
\value{
A Rel object.

A Rel object.

A character vector with the summary of the object.

The Rel object with the relationship informations.

A list with the relationship informations.

A data.frame with the relationship informations.
}
\description{
A Rel object is a list of special relationships
between individuals in the pedigree.
It is used to create a Pedigree object.
The minimal needed informations are \code{id1}, \code{id2} and \code{code}.
If a \code{family} is provided, the individuals \code{id} will be aggregated
to the \code{family} character to ensure the uniqueness of the \code{id}.

Constructor for the Rel class.

Compute the summary of a Rel object

Convert the Rel object to a data.frame
and print it with its summary.

Convert a Rel object to a list

Convert a Rel object to a data.frame
}
\section{Slots}{

\describe{
\item{\code{id1}}{A character vector with the id of the first individual.}

\item{\code{id2}}{A character vector with the id of the second individual.}

\item{\code{code}}{An ordered factor vector with the code of the special
relationship. (i.e. \verb{MZ twin} < \verb{DZ twin} < \verb{UZ twin} < \code{Spouse}).}

\item{\code{family}}{A character vector with the family of the individuals.}
}}

\seealso{
\code{\link[=Pedigree]{Pedigree()}}

\code{\link[=Pedigree]{Pedigree()}}
}
