% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleped}
\alias{sampleped}
\title{samplepedigree data}
\format{
A data frame with 55 observations, one line per subject, on the
following 7 variables.
\itemize{
\item \code{family} family identifier
\item \code{id} subject identifier
\item \code{dadid} identifier of the father, if the father is part of the
data set; zero otherwise
\item \code{momid} identifier of the mother, if the mother is part of the
data set; zero otherwise
\item \code{sex} 1 for male or 2 for F
\item \code{affected} 1 or 0
\item \code{available} 1 or 0
}
}
\usage{
data("sampleped")
}
\description{
Small sample pedigree data set.
}
\details{
This is a small fictive pedigree data set, with 55
individuals in 2 families.
The aim was to create a data set with a variety of pedigree structures.
}
\examples{
data("sampleped")
ped <- pedigree(sampleped)
summary(ped)
#plot(ped)
}
\keyword{datasets}
