% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Scale-class}
\alias{Scale-class}
\title{S4 class to represent the scales of a Pedigree.}
\value{
A Scale object.
}
\description{
A scales object is a list of two data.frame used
to represent the affection (filling) and the availability
(border) status of the individuals in the pedigree plot.
}
\section{Slots}{

\describe{
\item{\code{fill}}{A data.frame with the informations for the affection status.
The columns needed are:
\itemize{
\item 'order': the order of the affection to be used
\item 'column_values': name of the column containing the raw values in the
Ped object
\item 'column_mods': name of the column containing the mods of the transformed
values in the Ped object
\item 'mods': all the different mods
\item 'labels': the corresponding labels of each mods
\item 'affected': a logical value indicating if the mod correspond to an affected
individuals
\item 'fill': the color to use for this mods
\item 'density': the density of the shading
\item 'angle': the angle of the shading
}}

\item{\code{border}}{A data.frame with the informations for the availability status.
The columns needed are:
\itemize{
\item 'column_values': name of the column containing the raw values in the
Ped object
\item 'column_mods': name of the column containing the mods of the transformed
values in the Ped object
\item 'mods': all the different mods
\item 'labels': the corresponding labels of each mods
\item 'border': the color to use for this mods
}}
}}

\seealso{
\code{\link[=Pedigree]{Pedigree()}}
}
