% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructor.R, R/AllGeneric.R
\docType{methods}
\name{Scales}
\alias{Scales}
\alias{Scales,data.frame,data.frame-method}
\alias{Scales,data.frame,data.frame}
\alias{Scales,missing,missing-method}
\alias{Scales,missing,missing}
\alias{as.list,Scales-method}
\title{Create a Scales object}
\usage{
Scales(fill, border)

\S4method{Scales}{data.frame,data.frame}(fill, border)

\S4method{Scales}{missing,missing}(fill, border)

\S4method{as.list}{Scales}(x)
}
\arguments{
\item{fill}{A data.frame with the informations for the affection status.
The columns needed are:
\itemize{
\item 'order': the order of the affection to be used
\item 'column_values': name of the column containing the raw values in the
Ped object
\item 'column_mods': name of the column containing the mods of the transformed
affection in the Ped object
\item 'mods': all the different mods
\item 'labels': the corresponding labels of each mods
\item 'affected': a logical value indicating if the mod correspond to an affected
individuals
\item 'fill': the color to use for this mods
\item 'density': the density of the shading
\item 'angle': the angle of the shading
}}

\item{border}{A data.frame with the informations for the availability status.
The columns needed are:
\itemize{
\item 'column': name of the column containing the mods to use in the
Ped object
\item 'mods': all the different mods
\item 'labels': the corresponding labels of each mods
\item 'border': the color to use for this mods
}}

\item{from}{A Scales object.}
}
\value{
A Scales object.

A list with the hints informations.
}
\description{
Create a Scales object from two data.frame.
The first one is used to represent the affection status of the individuals
and therefore the filling of the individuals in the pedigree plot.
The second one is used to represent the availability status of the
individuals and therefore the border color of the individuals in the
pedigree plot.

Convert a Scales object to a list
}
\examples{
Scales(
   fill = data.frame(
      order = 1,
      column_values = "affected",
      column_mods = "affected_mod",
      mods = c(0, 1),
      labels = c("unaffected", "affected"),
      affected = c(FALSE, TRUE),
      fill = c("white", "red"),
      density = c(NA, 20),
      angle = c(NA, 45)
   ),
   border = data.frame(
      column = "avail",
      mods = c(0, 1),
      labels = c("not available", "available"),
      border = c("black", "blue")
   )
)
Scales()
}
\seealso{
\code{\link[=Pedigree]{Pedigree()}}
}
