% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{set_affected}
\alias{set_affected}
\title{Generate the affected matrix}
\usage{
set_affected(affected, n, angle, density)
}
\arguments{
\item{affected}{a vector or matrix of 0/1 values}

\item{n}{number of subjects}

\item{angle}{angle of shading}

\item{density}{density of shading}
}
\value{
a matrix of 0/1 values
}
\description{
Doc: still part of setup/data affected is a 0/1 matrix of any marker
data.  It may be attached to the pedigree or added here.  It can be a
vector of length [\link{n}] or a matrix with [\link{n}] rows. If not present,
the default is to plot open symbols without shading or color
If affected is a matrix, then the shading and/or density value for
ith column is taken from the ith element of the angle/density
arguments.
For purposes within the plot method, NA values in 'affected'
are coded to -1, and plotted as a question mark (?) in the plot
symbol region for that affected status
}
