% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{set_plot_area}
\alias{set_plot_area}
\title{Set plotting area}
\usage{
set_plot_area(cex, id, maxlev, xrange, symbolsize, ...)
}
\arguments{
\item{cex}{character expansion of the text}

\item{id}{A vector of each subjects identifiers}

\item{maxlev}{maximum level}

\item{xrange}{range of x values}

\item{symbolsize}{size of the symbols}

\item{...}{other arguments passed to \code{\link[=par]{par()}}}
}
\value{
a list of user coordinates, old par, box width, box height,
label height and leg height
}
\description{
Set plotting area
}
\keyword{internal}
