% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{shift}
\alias{shift}
\title{Routine to shift set of siblings to the left or right}
\usage{
shift(id, sibs, goleft, hint, twinrel, twinset)
}
\arguments{
\item{id}{The id of the subject to be shifted}

\item{sibs}{The ids of the siblings}

\item{goleft}{If TRUE, shift to the left, otherwise to the right}

\item{hint}{The current hint vector}

\item{twinrel}{The twin relationship matrix}

\item{twinset}{The twinset vector}
}
\value{
The updated hint vector
}
\description{
Routine to shift set of siblings to the left or right
}
\keyword{internal}
