% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{show,Ped-method}
\alias{show,Ped-method}
\title{Show function of Ped object}
\usage{
\S4method{show}{Ped}(object)
}
\arguments{
\item{object}{A Ped object.}
}
\value{
The Ped object with the individuals informations.
}
\description{
Convert the Ped object to a data.frame
and print it with its summary.
}
