% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\name{show,Pedigree-method}
\alias{show,Pedigree-method}
\title{Pedigree show method.}
\usage{
\S4method{show}{Pedigree}(object)
}
\arguments{
\item{object}{A Pedigree object.}
}
\value{
A character vector with the informations about the object.
}
\description{
Pedigree show method.
}
