% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\name{[,Pedigree,missing,ANY,ANY-method}
\alias{[,Pedigree,missing,ANY,ANY-method}
\title{Extract parts of a Pedigree object}
\usage{
\S4method{[}{Pedigree,missing,ANY,ANY}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{A Pedigree object.}

\item{i}{A vector of individuals id or a vector of index.}

\item{j}{A vector of columns names.}

\item{drop}{A logical value indicating if the dimensions should be dropped.}
}
\value{
A Pedigree object subsetted.
}
\description{
Extract parts of a Pedigree object
}
