% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\name{[[,Pedigree,ANY,missing-method}
\alias{[[,Pedigree,ANY,missing-method}
\title{Extract parts of a Pedigree object}
\usage{
\S4method{[[}{Pedigree,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A Pedigree object.}

\item{i}{A vector of individuals id or a vector of index.}

\item{j}{A vector of columns names.}

\item{...}{Other arguments.}

\item{drop}{A logical value indicating if the dimensions should be dropped.}
}
\value{
The slot \code{i} present in the pedigree object.
}
\description{
Extract parts of a Pedigree object
}
