% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\name{[[<-,Pedigree,ANY,missing-method}
\alias{[[<-,Pedigree,ANY,missing-method}
\title{Replace parts of a Pedigree object}
\usage{
\S4method{[[}{Pedigree,ANY,missing}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A Pedigree object.}

\item{i}{A vector of individuals id or a vector of index.}

\item{j}{A vector of columns names.}

\item{...}{Other arguments.}

\item{value}{A vector of values to replace.}
}
\value{
The pedigree object with the slot \code{i} replaced by \code{value}.
}
\description{
Replace parts of a Pedigree object
}
