% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{summary,Ped-method}
\alias{summary,Ped-method}
\title{Summary function of Ped object}
\usage{
\S4method{summary}{Ped}(object)
}
\arguments{
\item{object}{A Ped object.}
}
\value{
A character vector with the summary of the object.
}
\description{
Compute the summary of a Ped object
}
