% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\name{summary,Pedigree-method}
\alias{summary,Pedigree-method}
\title{Pedigree summary method.}
\usage{
\S4method{summary}{Pedigree}(object)
}
\arguments{
\item{object}{A Pedigree object.}
}
\value{
A character vector with the summary of the object.
}
\description{
Pedigree summary method.
}
