% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim a pedigree}
\usage{
trim(ped, id_rm, missid = "0")
}
\arguments{
\item{ped}{Pedigree object}

\item{id_rm}{Vector of ids to remove}

\item{missid}{Missing id to replace removed subjects}
}
\value{
A Pedigree object with the subjects removed
}
\description{
Carries out the removal of the subjects identified from a pedigree object.
}
