% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_famid.R
\name{upd_famid_id}
\alias{upd_famid_id}
\title{Update family prefix in individuals id}
\usage{
upd_famid_id(obj, famid, ...)
}
\arguments{
\item{obj}{A character vector of individual ids}

\item{famid}{A character vector with the family identifiers of the
individuals. If provide, will be aggregated to the individuals
identifiers separated by an underscore.}
}
\value{
A character vector of individual ids with family prefix
}
\description{
Update family prefix in individuals id
}
