% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_inds.R
\docType{methods}
\name{useful_inds}
\alias{useful_inds}
\alias{useful_inds,character-method}
\alias{useful_inds,character}
\alias{useful_inds,data.frame-method}
\alias{useful_inds,data.frame}
\alias{useful_inds,Pedigree-method}
\alias{useful_inds,Pedigree}
\title{Compute the usefulness of individuals}
\usage{
useful_inds(obj, ...)

\S4method{useful_inds}{character}(
  obj,
  dadid,
  momid,
  sex,
  avail,
  affected,
  num_child_tot,
  informative = "AvAf",
  keep_infos = FALSE,
  missid = "0"
)

\S4method{useful_inds}{data.frame}(obj, informative = "AvAf", keep_infos = FALSE, missid = "0")

\S4method{useful_inds}{Pedigree}(obj, reset = FALSE, ...)
}
\arguments{
\item{obj}{A dataframe, a Pedigree object or a character vector of ids}

\item{...}{Other arguments passed to methods.}

\item{dadid}{The id of the father}

\item{momid}{The id of the mother}

\item{sex}{A vector with the gender information}

\item{avail}{A vector of individuals availability (0, 1 or NA)}

\item{affected}{A vector of individuals affected status (0, 1 or NA)}

\item{num_child_tot}{A numeric vector of the number of children of each
individuals}

\item{informative}{Informative individuals selection can take 5 values:
'AvAf' (available and affected),
'AvOrAf' (available or affected),
'Av' (available only),
'Af' (affected only),
'All' (all individuals)
or a numeric/character vector of individuals id
or a boolean}

\item{keep_infos}{Boolean to indicate if individuals with unknown status
but available or reverse should be kept}

\item{missid}{A string with the missing id}

\item{reset}{Boolean to indicate if the \code{useful} column should be reset}
}
\description{
Check for usefulness of individuals
}
\details{
Check for the informativeness of the individuals based on the
informative parameter given, the number of children and the usefulness
of their parents. A \code{useful} column is added to the dataframe with the
usefulness of the individual. This boolean is hereditary.
}
