% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_to_factor}
\alias{var_to_factor}
\title{Variable to factor}
\usage{
var_to_factor(var, threshold = NULL)
}
\arguments{
\item{var}{The variable to be transformed}

\item{thresholds}{The thresholds to be used to separate the variable}
}
\value{
a factor vector containing the transformed variable
}
\description{
Transform a variable into a factor
}
\details{
Transform a numeric or string variable into a factor.
In the case of a numerical variable the threshold is used to separate
the variable using the `cut` function.
}
\examples{
var <- runif(10)
var_to_factor(var, threshold = 0.5)

}
