% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vect_to_binary}
\alias{vect_to_binary}
\title{Transform a vector variable to binary vector}
\usage{
vect_to_binary(vect, logical = FALSE)
}
\arguments{
\item{vect}{A character, factor, logical or numeric vector corresponding to
a binary variable (i.e. 0 or 1).
The following values are recognized:
\itemize{
\item character() or factor() : "TRUE", "FALSE", "0", "1", "NA" will be
respectively transformed to 1, 0, 0, 1, NA.
Spaces and case are ignored.
All other values will be transformed to NA.
\item numeric() : 0 and 1 are kept, all other values are transformed to NA.
\item logical() : TRUE and FALSE are tansformed to 1 and 0.
}}
}
\value{
numeric binary vector of the same size as \code{vect} with 0 and 1
}
\description{
Transform a vector variable to binary vector
}
\examples{
vect_to_binary(
   c(0, 1, 2, 3.6, "TRUE", "FALSE", "0", "1", "NA", "B", TRUE, FALSE, NA)
)
}
