% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Generate plotting information for a Pedigree}
\usage{
align(
  ped,
  packed = TRUE,
  width = 10,
  align = TRUE,
  hints = ped$hints,
  missid = "0"
)
}
\arguments{
\item{ped}{A Pedigree object}

\item{packed}{Should the Pedigree be compressed, i.e., allow diagonal
lines connecting parents to children in order to have a smaller overall
width for the plot.}

\item{width}{for a packed output, the minimum width of the plot, in
inches.}

\item{align}{for a packed Pedigree, align children under parents \code{TRUE},
to the extent possible given the page width, or align to to the left
margin \code{FALSE}.
This argument can be a two element vector, giving the alignment
parameters, or a logical value.
If \code{TRUE}, the default is \code{c(1.5, 2)}, or numeric the routine
\code{alignped4()} will be called.}

\item{hints}{Plotting hints for the Pedigree.
This is a list with components \code{order} and \code{spouse}, the second one
is optional.
\itemize{
\item \strong{order} is a numeric vector with one element per subject in the
Pedigree.  It determines the relative order of subjects within a sibship, as
well as the relative order of processing for the founder couples. (For this
latter, the female founders are ordered as though they were sisters).
\item \strong{spouse} is a matrix with one row per hinted marriage, usually
only a few marriages in a pedigree will need an added hint, for instance
reverse the plot order of a husband/wife pair. Each row contains the
index of the left spouse, the right hand spouse, and the anchor
(i.e : \code{1} = left, \code{2} = right, \code{0} = either).
Children will preferentially appear under the parents of the anchored
spouse.
}}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the Pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{"0"}.}
}
\value{
A list with components
\itemize{
\item n A vector giving the number of subjects on each horizonal level of the
plot
\item nid A matrix with one row for each level, giving the numeric id of
each subject plotted.
(A value of \code{17} means the 17th subject in the Pedigree).
\item pos A matrix giving the horizontal position of each plot point
\item fam A matrix giving the family id of each plot point.
A value of \code{3} would mean that the two subjects in positions 3 and 4,
in the row above, are this subject's parents.
\item spouse A matrix with values
\itemize{
\item \code{0} = not a spouse
\item \code{1} = subject plotted to the immediate right is a spouse
\item \code{2} = subject plotted to the immediate right is an inbred spouse
}
\item twins Optional matrix which will only be present if the Pedigree
contains twins :
\itemize{
\item \code{0} = not a twin
\item \code{1} = sibling to the right is a monozygotic twin
\item \code{2} = sibling to the right is a dizygotic twin
\item \code{3} = sibling to the right is a twin of unknown zygosity
}
}
}
\description{
Given a Pedigree, this function creates helper matrices that describe the
layout of a plot of the Pedigree.
}
\details{
This is an internal routine, used almost exclusively by
\code{ped_to_plotdf()}.  The subservient functions \code{auto_hint()},
\code{alignped1()}, \code{alignped2()},
\code{alignped3()}, and \code{alignped4()}
contain the bulk of the computation.
If the \strong{hints} are missing the \code{auto_hint()} routine is called to
supply an initial guess.
If multiple families are present in the Pedigree, this routine is called
once for each family, and the results are combined in the list returned.
For more information you can read the associated vignette:align
\code{vignette("alignment_details")}.
}
\examples{
data(sampleped)
ped <- Pedigree(sampleped)
align(ped)

}
\seealso{
\code{\link[=alignped1]{alignped1()}},
\code{\link[=alignped2]{alignped2()}},
\code{\link[=alignped3]{alignped3()}},
\code{\link[=alignped4]{alignped4()}},
\code{\link[=auto_hint]{auto_hint()}}
}
