% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Generate plotting information for a pedigree}
\usage{
align(obj, ...)
}
\arguments{
\item{ped}{A pedigree object}

\item{packed}{Should the pedigree be compressed, i.e., to allow diagonal
lines connecting parents to children in order to have a smaller overall
width for the plot.}

\item{hints}{Plotting hints for the pedigree.
This is a list with components `order` and `spouse`,
the second one is optional. If the hints are missing the `auto_hint`
routine is called to supply an initial guess.
The order component is a numeric vector with one element per subject in the
pedigree.  It determines the relative order of subjects within a sibship, as
well as the relative order of processing for the founder couples. (For this
latter, the female founders are ordered as though they were sisters). The
spouse component is a matrix with one row per hinted marriage, usually only
a few marriages in a pedigree will need an added hint, for instance reverse
the plot order of a husband/wife pair. Each row contains the index of the
left spouse, the right hand spouse, and the anchor:
1 = left, 2 = right, 0 = either.
Children will preferentially appear under the parents of the anchored spouse.}

\item{width}{for a packed output, the minimum width}

\item{align}{for a packed pedigree, align children under parents (TRUE), to
the extent possible given the page width, or align to to the left margin
(FALSE). This argument can be a two element vector, giving the alignment
parameters, or a logical value.  If TRUE, the default is c(1.5, 2), or
numeric the routine `alignped4` will be called.}
}
\value{
a structure with components
## n
A vector giving the number of subjects
on each horizonal level of the plot
## nid
A matrix with one row for each level, giving the numeric id of
each subject plotted.
(An value of 17 means the 17th subject in the pedigree).
## pos
A matrix giving the horizontal position of each plot point
## fam
A matrix giving the family id of each plot point.
A value of '3' would mean that the two subjects in positions 3 and 4,
in the row above, are this subject's parents.
## spouse
A matrix with values
1 = subject plotted to the immediate right is a spouse
2 = subject plotted to the immediate right is an inbred spouse
0 = not a spouse
## twins
Optional matrix which will only be present if the pedigree
contains twins :
1 = sibling to the right is a monozygotic twin
2 = sibling to the right is a dizygotic twin
3 = sibling to the right is a twin of unknown zygosity
0 = not a twin
}
\description{
Given a pedigree, this function creates helper matrices that descibe the
layout of a plot of the pedigree.
}
\details{
This is an internal routine, used almost exclusively by
`plot.pedigree`.  The subservient functions `alignped1`,
`alignped2`, `alignped3`, and `alignped4` contain the bulk of
the computation.
}
\examples{
data(sample.ped)
ped <- with(sample.ped, pedigree(id, father, mother, sex, affected))
align(ped)

}
\seealso{
`plot.pedigree`, `auto_hint`
}
\keyword{dplot}
