% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignped3.R
\name{alignped3}
\alias{alignped3}
\title{Third routine alignement}
\usage{
alignped3(x1, x2, packed, space = 1)
}
\arguments{
\item{space}{}
}
\value{
A list of element containing:
## n
A vector giving the number of subjects
on each horizonal level of the plot
## nid
A matrix with one row for each level, giving the numeric id of
each subject plotted.
(An value of 17 means the 17th subject in the pedigree).
## pos
A matrix giving the horizontal position of each plot point
## fam
A matrix giving the family id of each plot point.
A value of '3' would mean that the two subjects in positions 3 and 4,
in the row above, are this subject's parents.
}
\description{
This is the third of the four co-routines.
}
\details{
This co-routine merges two pedigree trees which are side by side into a
single object. The primary special case is when the rightmost person in
the left tree is the same as the leftmost person in the right tree; we
need not plot two copies of the same person side by side.
(When initializing the output structures do not worry about this,
there is no harm if they are a column bigger than finally needed.)
Beyond that the work is simple bookkeeping.

## Slide
For the unpacked case, which is the traditional way to draw
a pedigree when we can assume the paper is infinitely wide, all parents are
centered over their children. In this case we think if the two trees to be
merged as solid blocks. On input they both have a left margin of 0.
Compute how far over we have to slide the right tree.
## Merge
Now merge the two trees. Start at the top level and work down.
}
\examples{
data(sample.ped)
ped <- with(sample.ped, pedigree(id, father, mother, sex, affected))
align.pedigree(ped)

}
\seealso{
`plot.pedigree`, `autohint`
}
\keyword{dplot}
