% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{ancestor}
\alias{ancestor}
\title{Routine function to get ancestors of a subject}
\usage{
ancestor(idx, momx, dadx)
}
\arguments{
\item{idx}{Index of the subject}

\item{momx}{Indexs of the mothers}

\item{dadx}{Indexs of the fathers}
}
\value{
A vector of ancestor indexs
}
\description{
Given the index of one individual, this
function iterate through the mom and dad indexs to
list out all the ancestors of the said individual.
This function is use in the \code{align()} function to
identify which spouse pair has a common ancestor and
therefore if they need to be connected with a double line.
}
\seealso{
\code{\link[=align]{align()}}
}
