% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.pedigree.R
\name{as.data.frame.pedigree}
\alias{as.data.frame.pedigree}
\title{data.frame from a pedigree object}
\usage{
as.data.frame.pedigree(x, ...)
}
\arguments{
\item{x}{pedigree object}
\item{...}{additional arguments passed to internal methods}
}
\value{
A data.frame with the data necessary to re-create the pedigree,
minus special relationships.
}
\author{
Jason Sinnwell
}
\description{
Extract the internal data from a pedigree object into a data.frame
}
\seealso{
\code{\link{pedigree}}
}
