% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_size.R
\docType{methods}
\name{bit_size}
\alias{bit_size}
\alias{bit_size,character_OR_integer-method}
\alias{bit_size,character}
\alias{bit_size,Pedigree-method}
\alias{bit_size,Pedigree}
\alias{bit_size,Ped-method}
\alias{bit_size,Ped}
\title{Get Pedigree bit_size}
\usage{
bit_size(obj, ...)

\S4method{bit_size}{character_OR_integer}(obj, momid, missid = NA_character_)

\S4method{bit_size}{Pedigree}(obj, missid = NA_character_)

\S4method{bit_size}{Ped}(obj, missid = NA_character_)
}
\arguments{
\item{obj}{A Pedigree object or a vector of fathers identifierss}

\item{...}{Additional arguments passed to methods}
}
\value{
A list with the following components:
\itemize{
\item bit_size The bit_size of input Pedigree
\item nFounder The number of founders in the Pedigree
\item nNonFounder The number of nonfounders in the Pedigree
}
}
\description{
Calculate Pedigree bit_size, defined as :

\eqn{
  2 \times NbNonFounders - NbFounders
}
}
\details{
This is a utility function used in \code{shrink()}
to calculate the bit_size of a Pedigree.
}
\examples{
data(sampleped)
ped <- Pedigree(sampleped)
bit_size(ped)
}
\seealso{
\code{\link[=shrink]{shrink()}}
}
\keyword{internal}
