% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_size.R
\docType{methods}
\name{bit_size}
\alias{bit_size}
\alias{bit_size,character-method}
\alias{bit_size,character}
\alias{bit_size,Pedigree-method}
\alias{bit_size,Pedigree}
\title{Get pedigree bit_size}
\usage{
bit_size(obj, ...)

\S4method{bit_size}{character}(obj, momid, missid = "0")

\S4method{bit_size}{Pedigree}(obj, missid = "0")
}
\arguments{
\item{obj}{A pedigree object or a vector of fathers identifierss}

\item{...}{Additional arguments passed to methods}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{"0"}.}
}
\value{
A list with the following components:
\itemize{
\item bit_size The bit_size of input pedigree
\item nFounder The number of founders in the pedigree
\item nNonFounder The number of nonfounders in the pedigree
}
}
\description{
Calculate pedigree bit_size, defined as :

\eqn{
  2 \times NbNonFounders - NbFounders
}
}
\details{
This is a utility function used in \code{shrink()}
to calculate the bit_size of a pedigree.
}
\examples{
data(sampleped)
ped <- pedigree(sampleped)
bit_size(ped)
}
\seealso{
\code{\link[=shrink]{shrink()}}
}
\keyword{internal}
