% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_size.R
\name{bit_size}
\alias{bit_size}
\title{Get pedigree bit_size}
\usage{
bit_size(obj, ...)
}
\arguments{
\item{dadid}{Vector of fathers ids}

\item{momid}{Vector of mothers ids}

\item{missid}{Character defining the missing ids}
}
\value{
A list with the following components:
## bit_size
The bit_size of input pedigree
## nFounder
The number of founders in the pedigree
## nNonFounder
The number of nonfounders in the pedgiree
}
\description{
Calculate pedigree bit_size, defined as 2 * # NonFounders - # Founders
}
\details{
This is a utility function used in `pedigree.shrink()`
}
\seealso{
`pedigree.shrink`
}
