% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{check_values}
\alias{check_values}
\title{Check values in a slot}
\usage{
check_values(val, ref, present = TRUE)
}
\description{
Check if the all the values in a slot are in a vector of values.
}
\details{
$param object An object.
$param slot A slot of the object.
$param column A column of the slot.
$param values A vector of values to check.
$param present A logical value indicating if the values should be present
or not

$return A character vector with the errors if any.
}
