% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllValidity.R
\name{check_values}
\alias{check_values}
\title{Check values in a slot}
\usage{
check_values(val, ref, name = NULL, present = TRUE)
}
\arguments{
\item{val}{A vector of values to check.}

\item{ref}{A vector of reference values.}

\item{name}{A character vector with the name of the values to check.}

\item{present}{A logical value indicating if the values should be present
or not}
}
\value{
A character vector with the errors if any.
}
\description{
Check if the all the values in a slot are in a vector of values.
}
\keyword{internal}
