% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\docType{methods}
\name{descendants}
\alias{descendants}
\alias{descendants,character_OR_integer,character_OR_integer-method}
\alias{descendants,character_OR_integer}
\alias{descendants,character_OR_integer,Pedigree-method}
\alias{descendants,Pedigree}
\alias{descendants,character_OR_integer,Ped-method}
\alias{descendants,Ped}
\title{Find all the descendants}
\usage{
descendants(idlist, obj, ...)

\S4method{descendants}{character_OR_integer,character_OR_integer}(idlist, obj, dadid, momid)

\S4method{descendants}{character_OR_integer,Pedigree}(idlist, obj)

\S4method{descendants}{character_OR_integer,Ped}(idlist, obj)
}
\arguments{
\item{idlist}{List of individuals identifiers to be considered}

\item{obj}{A Pedigree object}

\item{...}{Additional arguments passed to methods}
}
\value{
Vector of all descendants of the individuals in idlist.
The list is not ordered.
}
\description{
Find all the descendants of a particular list of individuals
given a Pedigree
}
\examples{
data("sampleped")
ped <- Pedigree(sampleped)
descendants(c("1_101", "2_208"), ped)
}
\keyword{internal}
