% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{descendants}
\alias{descendants}
\title{Find all the descendants}
\usage{
descendants(idlist, obj, ...)
}
\arguments{
\item{idlist}{List of individuals id to be considered}

\item{id}{Individual id of the pedigree}

\item{dadid}{Father id of the pedigree}

\item{momid}{Mother id of the pedigree}
}
\value{
List of all the descendants of the individuals in idlist.
The list is not ordered.
}
\description{
Find all the descendants of a particular list of individuals
given a pedigree
}
\examples{
data("sampleped")
ped <- pedigree(sampleped)
descendants(c("1_101", "2_208"), ped)
}
