% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\docType{methods}
\name{descendants}
\alias{descendants}
\alias{descendants,character,character-method}
\alias{descendants,character}
\alias{descendants,character,Pedigree-method}
\alias{descendants,Pedigree}
\title{Find all the descendants}
\usage{
descendants(idlist, obj, ...)

\S4method{descendants}{character,character}(idlist, obj, dadid, momid)

\S4method{descendants}{character,Pedigree}(idlist, obj)
}
\arguments{
\item{idlist}{List of individuals identifiers to be considered}

\item{obj}{A pedigree object or a vector of subject identifiers.}

\item{...}{Additional arguments passed to methods}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}
}
\value{
Vector of all descendants of the individuals in idlist.
The list is not ordered.
}
\description{
Find all the descendants of a particular list of individuals
given a Pedigree
}
\examples{
data("sampleped")
ped <- Pedigree(sampleped)
descendants(c("1_101", "2_208"), ped)
}
\keyword{internal}
