% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_arc}
\alias{draw_arc}
\title{Draw arcs for multiple instances of a subject}
\usage{
draw_arc(x, y, p, ggplot_gen = FALSE)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{p}{ggplot object}

\item{ggplot_gen}{logical, if TRUE add the arcs to the ggplot object}
}
\value{
Plot the arcs or add it to a ggplot object
}
\description{
Draw arcs for multiple instances of a subject
}
