% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_arc}
\alias{draw_arc}
\title{Draw arcs}
\usage{
draw_arc(x0, y0, x1, y1, p, ggplot_gen = FALSE, lwd = 1, col = "black")
}
\arguments{
\item{x0}{x coordinate of the first point}

\item{y0}{y coordinate of the first point}

\item{x1}{x coordinate of the second point}

\item{y1}{y coordinate of the second point}

\item{p}{ggplot object}

\item{ggplot_gen}{If TRUE add the segments to the ggplot object}

\item{lwd}{Line width}

\item{col}{Line color}
}
\value{
Plot the arcs to the current device
or add it to a ggplot object
}
\description{
Draw arcs
}
\keyword{Pedigree-plot}
\keyword{internal,}
