% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_polygon}
\alias{draw_polygon}
\title{Draw a polygon}
\usage{
draw_polygon(
  x,
  y,
  p,
  ggplot_gen = FALSE,
  fill = "grey",
  border = NULL,
  density = NULL,
  angle = 45
)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{p}{ggplot object}

\item{ggplot_gen}{If TRUE add the segments to the ggplot object}

\item{fill}{Fill color}

\item{border}{Border color}

\item{density}{Density of shading}

\item{angle}{Angle of shading}
}
\value{
Plot the polygon  to the current device
or add it to a ggplot object
}
\description{
Draw a polygon
}
\keyword{Pedigree-plot}
\keyword{internal,}
