% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_text}
\alias{draw_text}
\title{Draw text for a pedigree}
\usage{
draw_text(x, y, label, p, ggplot_gen = FALSE, cex = 1, col = NULL)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{label}{text to be displayed}

\item{p}{ggplot object}

\item{ggplot_gen}{logical, if TRUE add the text to the ggplot object}
}
\value{
Plot the text or add it to a ggplot object
}
\description{
Draw text for a pedigree
}
