% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_text}
\alias{draw_text}
\title{Draw texts}
\usage{
draw_text(
  x,
  y,
  label,
  p,
  ggplot_gen = FALSE,
  cex = 1,
  col = NULL,
  adjx = 0,
  adjy = 0
)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{label}{Text to be displayed}

\item{p}{ggplot object}

\item{ggplot_gen}{If TRUE add the segments to the ggplot object}

\item{cex}{Character expansion of the text}

\item{col}{Text color}

\item{adjx}{x adjustment}

\item{adjy}{y adjustment}
}
\value{
Plot the text to the current device
or add it to a ggplot object
}
\description{
Draw texts
}
\keyword{Pedigree-plot}
\keyword{internal,}
