% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{duporder}
\alias{duporder}
\title{Routine to find the duplicate pairs of a subject}
\usage{
duporder(idlist, plist, lev, ped)
}
\arguments{
\item{idlist}{The list of ids}

\item{plist}{The alignment structure}

\item{lev}{The generation level of the subject}

\item{ped}{A pedigre object}
}
\value{
A matrix of duplicate pairs
}
\description{
Routine to find the duplicate pairs of a subject
}
\details{
This routine is used by \code{auto_hint()}.
It finds the duplicate pairs of a subject and returns them in
the order they should be plotted.
}
\seealso{
\code{\link[=auto_hint]{auto_hint()}}
}
