% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{duporder}
\alias{duporder}
\title{Find the duplicate pairs of a subject}
\usage{
duporder(idlist, plist, lev, obj)
}
\arguments{
\item{idlist}{List of individuals identifiers to be considered}

\item{plist}{The alignment structure representing the Pedigree layout.
See \code{\link[=align]{align()}} for details.}

\item{lev}{The generation level of the subject}

\item{obj}{A Pedigree object}
}
\value{
A matrix of duplicate pairs
}
\description{
Find the duplicate pairs of a subject
}
\details{
This routine is used by \code{auto_hint()}.
It finds the duplicate pairs of a subject and returns them in
the order they should be plotted.
}
\seealso{
\code{\link[=auto_hint]{auto_hint()}}
}
\keyword{auto_hint}
\keyword{internal,}
