% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\name{show,Pedigree-method}
\alias{show,Pedigree-method}
\alias{summary,Pedigree-method}
\alias{sub_sel_hints}
\alias{sub_sel_hints,Pedigree-method}
\title{Pedigree show method.}
\usage{
\S4method{show}{Pedigree}(object)

\S4method{summary}{Pedigree}(object)

sub_sel_hints(hints, index)
}
\arguments{
\item{object}{A Pedigree object.}

\item{hints}{A list of hints}

\item{index}{A vector of index}
}
\value{
A character vector with the informations about the object.

A character vector with the summary of the object.

A list of hints subsetted
}
\description{
Pedigree show method.

Pedigree summary method.

Subset the hints list based on the index given
}
\keyword{internal}
