% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllAccessors.R, R/AllGeneric.R
\name{ped}
\alias{ped}
\alias{ped,Pedigree-method}
\alias{mcols,Pedigree-method}
\alias{rel}
\alias{rel,Pedigree-method}
\alias{scales}
\alias{scales,Pedigree-method}
\alias{fill}
\alias{fill,Pedigree-method}
\alias{border}
\alias{border,Pedigree-method}
\alias{hints}
\alias{hints,Pedigree-method}
\alias{spouse}
\alias{spouse,Pedigree-method}
\alias{subset,Rel-method}
\alias{subset,Hints-method}
\alias{subset_hints,Hints-method}
\alias{show,Pedigree-method}
\alias{summary,Pedigree-method}
\alias{as.list,Pedigree-method}
\alias{subset,Pedigree-method}
\title{Pedigree ped accessors}
\usage{
ped(object, slot)

\S4method{mcols}{Pedigree}(x)

rel(object, slot)

scales(object)

fill(object)

border(object)

hints(object)

spouse(object)

\S4method{subset}{Rel}(x, idlist, keep = TRUE)

\S4method{subset}{Hints}(x, idlist, keep = TRUE)

\S4method{show}{Pedigree}(object)

\S4method{summary}{Pedigree}(object)

\S4method{as.list}{Pedigree}(x)

\S4method{subset}{Pedigree}(x, i, del_parents = FALSE, keep = TRUE)
}
\arguments{
\item{object}{A Pedigree object.}

\item{slot}{A slot in the Ped object of the Pedigree.}

\item{x}{A Pedigree object.}

\item{idlist}{A vector of identifiers to subset}

\item{i}{A vector of individuals id or a vector of index.}

\item{from}{A Pedigree object.}

\item{j}{A vector of columns names.}

\item{drop}{A logical value indicating if the dimensions should be dropped.}
}
\value{
The slot \code{ped} present in the Pedigree object.
or one of its slot.

The metadata present in the Pedigree object.

The slot \code{rel} present in the Pedigree object.

The slot \code{scales} present in the Pedigree object.

The slot \code{fill} present in the Pedigree object.

The slot \code{border} present in the Pedigree object.

The slot \code{hints} present in the Pedigree object.

The slot \code{spouse} present in the \code{Hints} slot of
a Pedigree object.

A Rel object subsetted.

A list of Hints object subsetted

A character vector with the informations about the object.

A character vector with the summary of the object.

A list with the individuals informations.

A Pedigree object subsetted.
}
\description{
Pedigree rel accessors

Pedigree scales accessors

Pedigree fill accessors

Pedigree border accessors

Pedigree hints accessors

Pedigree spouse accessors

Subset a Rel object based on the individuals
identifiers given.

Subset the Hints object based on the identifiers
given

Pedigree show method

Pedigree summary method.

Convert a Pedigree object to a list

Extract parts of a Pedigree object
}
\keyword{internal}
