% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllAccessors.R, R/AllGeneric.R
\name{ped}
\alias{ped}
\alias{ped,Pedigree-method}
\alias{mcols,Pedigree-method}
\alias{rel}
\alias{rel,Pedigree-method}
\alias{scales}
\alias{scales,Pedigree-method}
\alias{fill}
\alias{fill,Pedigree-method}
\alias{border}
\alias{border,Pedigree-method}
\alias{hints}
\alias{hints,Pedigree-method}
\alias{spouse}
\alias{spouse,Pedigree-method}
\alias{as.list,Ped-method}
\alias{as.data.frame,Ped-method}
\alias{show,Pedigree-method}
\alias{summary,Pedigree-method}
\alias{sub_sel_hints}
\alias{sub_sel_hints,Pedigree-method}
\alias{[,Pedigree,ANY,missing,ANY-method}
\title{Pedigree ped accessors}
\usage{
ped(object, slot)

\S4method{mcols}{Pedigree}(x)

rel(object, slot)

scales(object)

fill(object, slot)

border(object, slot)

hints(object)

spouse(object)

\S4method{as.list}{Ped}(x)

\S4method{as.data.frame}{Ped}(x)

\S4method{show}{Pedigree}(object)

\S4method{summary}{Pedigree}(object)

sub_sel_hints(hints, index)

\S4method{[}{Pedigree,ANY,missing,ANY}(x, i, j, drop = TRUE)
}
\arguments{
\item{object}{A Pedigree object.}

\item{slot}{A slot in the Ped object of the Pedigree.}

\item{x}{A Pedigree object.}

\item{hints}{A list of hints}

\item{index}{A vector of index}

\item{i}{A vector of individuals id or a vector of index.}

\item{j}{A vector of columns names.}

\item{drop}{A logical value indicating if the dimensions should be dropped.}

\item{from}{A Ped object.}
}
\value{
The slot \code{ped} present in the Pedigree object.
or one of its slot.

The metadata present in the Pedigree object.

The slot \code{rel} present in the Pedigree object.

The slot \code{scales} present in the Pedigree object.

The slot \code{fill} present in the Pedigree object.

The slot \code{border} present in the Pedigree object.

The slot \code{hints} present in the Pedigree object.

The slot \code{spouse} present in the \code{Hints} slot of
a Pedigree object.

A list with the individuals informations.
The metadata are put at the end.

A data.frame with the individuals informations.
The metadata are put at the end.

A character vector with the informations about the object.

A character vector with the summary of the object.

A list of hints subsetted

A Pedigree object subsetted.
}
\description{
Pedigree rel accessors

Pedigree scales accessors

Pedigree fill accessors

Pedigree border accessors

Pedigree hints accessors

Pedigree spouse accessors

Convert a Ped object to a list

Convert a Ped object to a data.frame

Pedigree show method

Pedigree summary method.

Subset the hints list based on the index given

Extract parts of a Pedigree object
}
\keyword{internal}
