% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_avail_affected.R
\name{find_avail_affected}
\alias{find_avail_affected}
\title{Find a single person to trim from a pedigree whose is available}
\usage{
find_avail_affected(ped, avail = ped$ped$avail, affstatus = NA)
}
\arguments{
\item{ped}{A pedigree objects, with id (subject ID), findex (father index),
mindex (mother index)}

\item{avail}{Vector of availability status (e.g., genotyped) 0/1 or
TRUE/FALSE}

\item{affstatus}{Vector of affection status 0/1 or TRUE/FALSE.}
}
\value{
A list is returned with the following components
\subsection{ped}{

Dataframe with trimmed subject removed
}

\subsection{idTrimmed}{

Vector of IDs of trimmed individuals
}

\subsection{isTrimmed}{

logical value indicating whether pedigree has been trimmed
}

\subsection{bit_size}{

Bit size of the trimmed pedigree
}
}
\description{
Finds one subject from among available non-parents with indicated affection
status.
}
\details{
When used within pedigree.shrink, this function is called with the first
affected indicator, if the affected item in the pedigree is a matrix of
multiple affected indicators.
}
\seealso{
\code{pedigree.shrink}
}
