% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_avail_noninform.R
\name{find_avail_noninform}
\alias{find_avail_noninform}
\title{Find uninformative but available subject}
\usage{
find_avail_noninform(obj, ...)
}
\arguments{
\item{obj}{A Pedigree object}

\item{...}{Other arguments to pass to the function}
}
\value{
Vector of subject ids who can be removed by having lowest
informativeness.
}
\description{
Find uninformative but available subject
}
\details{
Find subjects from a Pedigree who are available and uninformative

Identify subjects to remove from a Pedigree who are available but
non-informative.  This is the second step to remove subjects in
pedigree.shrink if the Pedigree does not meet the desired bit size.
}
\examples{
data(sampleped)
ped <- Pedigree(sampleped)
find_avail_noninform(ped)

}
\seealso{
\code{\link[=shrink]{shrink()}}
}
