% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{findsibs}
\alias{findsibs}
\title{Routine to find the siblings of a subject}
\usage{
findsibs(mypos, plist, lev)
}
\arguments{
\item{mypos}{The position of the subject}

\item{plist}{The alignment structure}

\item{lev}{The level of the subject}
}
\value{
The positions of the siblings
}
\description{
Routine to find the siblings of a subject
}
\keyword{internal}
