% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{findspouse}
\alias{findspouse}
\title{Routine to find the spouse of a subject}
\usage{
findspouse(idpos, plist, lev, ped)
}
\arguments{
\item{idpos}{The position of the subject}

\item{plist}{The alignment structure representing the pedigree layout.
For the differents matrices present in the list, each row represents a
level of the pedigree and each column a potential subject.
It contains the following components:
\itemize{
\item n Vector of the number of subjects per level
\item nid Matrix of the subjects indexes
\item pos Matrix of the subjects positions
\item fam Matrix of the siblings family identifiers
\item spouse Matrix of the left spouses \code{1} = spouse, \code{0} = not spouse,
\code{2} = inbred spouse.
}}

\item{lev}{The generation level of the subject}

\item{ped}{A pedigree object}
}
\value{
The position of the spouse
}
\description{
Routine to find the spouse of a subject
}
\details{
This routine is used by \code{auto_hint()}.
It finds the spouse of a subject.
}
\seealso{
\code{\link[=auto_hint]{auto_hint()}}
}
