% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{findspouse}
\alias{findspouse}
\title{Routine to find the spouse of a subject}
\usage{
findspouse(idpos, plist, lev, ped)
}
\arguments{
\item{idpos}{The position of the subject}

\item{plist}{The alignment structure}

\item{lev}{The generation level of the subject}

\item{ped}{A pedigre object}
}
\value{
The position of the spouse
}
\description{
Routine to find the spouse of a subject
}
\details{
This routine is used by \code{auto_hint()}.
It finds the spouse of a subject.
}
\seealso{
\code{\link[=auto_hint]{auto_hint()}}
}
