% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{findspouse}
\alias{findspouse}
\title{Find the spouse of a subject}
\usage{
findspouse(idpos, plist, lev, obj)
}
\arguments{
\item{idpos}{The position of the subject}

\item{plist}{The alignment structure representing the Pedigree layout.
See \code{\link[=align]{align()}} for details.}

\item{lev}{The generation level of the subject}

\item{obj}{A Pedigree object}
}
\value{
The position of the spouse
}
\description{
Find the spouse of a subject
}
\details{
This routine is used by \code{auto_hint()}.
It finds the spouse of a subject.
}
\seealso{
\code{\link[=auto_hint]{auto_hint()}}
}
\keyword{auto_hint}
\keyword{internal,}
