% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_aff_inds.R
\name{generate_aff_inds}
\alias{generate_aff_inds}
\title{Process the information for affection}
\usage{
generate_aff_inds(
  values,
  mods_aff = NULL,
  threshold = NULL,
  sup_thres_aff = NULL
)
}
\arguments{
\item{values}{Vector containing the values of the column to process.}

\item{mods_aff}{Vector of modality to consider as affected in the case
where the \code{values} is a factor.}

\item{threshold}{Numeric value separating the affected and healthy subject
in the case where the \code{values} is numeric.}

\item{sup_thres_aff}{Boolean defining if the affected individual are above
the threshold or not. If TRUE, the individuals will be considered affected
if the value of \code{values} is stricly above the \code{threshold}. If FALSE, the
individuals will be considered affected if the value is stricly under the
\code{threshold}.}
}
\value{
A dataframe with the \code{affected} column processed accordingly.
The different columns are:
\itemize{
\item \code{mods}: The different modalities of the column
\item \code{labels}: The labels of the different modalities
\item \code{affected}: The column processed to have only TRUE/FALSE values
}
}
\description{
Process the information for affection
}
\details{
Perform transformation uppon a column given as the one
containing affection status to get an \code{affected} column usable for
the rest of the script
}
\examples{
generate_aff_inds(c(1, 2, 3, 4, 5), threshold = 3, sup_thres_aff = TRUE)
generate_aff_inds(c("A", "B", "C", "A", "V", "B"), mods_aff = c("A", "B"))
}
