% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_colors.R
\name{generate_colors}
\alias{generate_colors}
\title{Process the colors based on affection and availability}
\usage{
generate_colors(obj, ...)
}
\arguments{
\item{df}{The dataframe containing the columns to process.}

\item{col_aff}{The name of the column containing the affection status.}

\item{mods_aff}{Vector of modality to consider as affected in the case
where the `col_aff` is a factor.}

\item{threshold}{Numeric value separating the affected and healthy subject
in the case where the `col_aff` is numeric.}

\item{sup_thres_aff}{Boolean defining if the affected individual are above
the threshold or not. If TRUE, the individuals will be considered affected
if the value of `col_aff` is stricly above the `threshold`. If FALSE, the
individuals will be considered affected if the value is stricly under the
`threshold`.}

\item{col_avail}{The name of the column containing the availability status.}

\item{keep_full_scale}{Boolean defining if the affection values need to
be set as a continuous scale or to a boolean.}

\item{breaks}{Number of breaks to use when using full scale with numeric
values. The same number of breaks will be done for values from affected
individuals and unaffected individuals.}

\item{colors_aff}{Set of increasing colors to use for the filling of the
affected individuls.}

\item{colors_unaff}{Set of increasing colors to use for the filling of the
unaffected individuls.}

\item{colors_avail}{Set of 2 colors to use for the box's border of an
individual. The first color will be used for available individual (avail
== 1) and the second for the unavailable individual (avail == 0).}
}
\value{
A list of two elements
- The processed dataframe with the `affected` and `avail` columns
processed accordingly
- A dataframe containing the description of each modality of the scale
}
\description{
Process the colors based on affection and availability
}
\details{
Perform transformation uppon a dataframe given to compute
the colors for the filling and the border of the individuals based
on the affection and availability status.
}
