% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{get_twin_rel}
\alias{get_twin_rel}
\title{Routine to get twin relationships}
\usage{
get_twin_rel(ped)
}
\value{
A list containing components
\enumerate{
\item \code{twinset} the set of twins
\item \code{twinrel} the twins relationships
\item \code{twinord} the order of the twins
}
}
\description{
Routine to get twin relationships
}
\details{
This routine function determine the twin relationships
in a Pedigree. It complete the missing twin relationships for
triplets, quads, etc. It also determine the order of the twins
in the Pedigree.
It is used by \code{auto_hint()}.
}
\seealso{
\code{\link[=auto_hint]{auto_hint()}}
}
\keyword{internal}
