% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_disconnected}
\alias{is_disconnected}
\title{Check wich individuals are disconnected}
\usage{
is_disconnected(id, dadid, momid)
}
\arguments{
\item{id}{A vector of each subjects identifiers}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}
}
\value{
A vector of boolean of the same size as \code{id}
with TRUE if the individual is disconnected and FALSE otherwise
}
\description{
Check which individuals are disconnected.
}
\details{
An individuals is considered disconnected if the kinship with
all the other individuals is 0.
}
