% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_disconnected}
\alias{is_disconnected}
\title{Are individuals disconnected}
\usage{
is_disconnected(id, dadid, momid)
}
\arguments{
\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{A vector containing for each subject, the identifiers of the
biologicals mothers.}
}
\value{
A vector of boolean of the same size as \strong{id}
with \code{TRUE} if the individual is disconnected and
\code{FALSE} otherwise
}
\description{
Check which individuals are disconnected.
}
\details{
An individuals is considered disconnected if the kinship with
all the other individuals is \code{0}.
}
\examples{
is_disconnected(
     c("1", "2", "3", "4", "5"),
     c("3", "3", NA, NA, NA),
     c("4", "4", NA, NA, NA)
)
}
\keyword{internal}
