% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_founder}
\alias{is_founder}
\title{Check wich individuals are founders}
\usage{
is_founder(momid, dadid, missid = "0")
}
\arguments{
\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the Pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{"0"}.}
}
\value{
A vector of boolean of the same size as \code{dadid} and \code{momid}
with TRUE if the individual has no parents (i.e is a founder) and FALSE
otherwise.
}
\description{
Check which individuals are founders.
}
\keyword{internal}
