% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_informative.R
\name{is_informative}
\alias{is_informative}
\title{Is informative}
\usage{
is_informative(df, informative = "AvAf")
}
\arguments{
\item{df}{Dataframe with individuals to select}

\item{informative}{Informative individuals selection can take 3 values:
'AvAf' (available and affected),
'AvOrAf' (available or affected),
'Av' (available only),
'Af' (affected only),
'All' (all individuals)
or a numeric vector of individuals id
or a boolean}
}
\description{
Select the ids of the informative individuals.
}
\details{
Depending on the informative parameter, the function will extract
the ids of the informative individuals. In the case of a numeric vector,
the function will return the same vector. In the case of a boolean, the
function will return the ids of the individuals if TRUE, NA otherwise.
In the case of a string, the function will return the ids of the
corresponding informative individuals based on the avail and affected
columns.
}
