% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_informative.R
\docType{methods}
\name{is_informative}
\alias{is_informative}
\alias{is_informative,character-method}
\alias{is_informative,character}
\alias{is_informative,Pedigree-method}
\alias{is_informative,Pedigree}
\title{Is informative}
\usage{
is_informative(obj, ...)

\S4method{is_informative}{character}(
  obj,
  avail,
  affected,
  informative = "AvAf",
  missid = NA_character_
)

\S4method{is_informative}{Pedigree}(
  obj,
  col_aff = NULL,
  informative = "AvAf",
  missid = NA_character_,
  reset = FALSE
)
}
\arguments{
\item{obj}{A character vector with the id of the individuals or a
\code{data.frame} with all the informations in corresponding columns.}

\item{...}{Additional arguments passed to methods}

\item{avail}{A numeric vector with the availability status of the
individuals (i.e. \code{0} = not available, \code{1} = available, \code{NA} = unknown).}

\item{affected}{A numeric vector with the affection status of the
individuals (i.e. \code{0} = unaffected, \code{1} = affected, \code{NA} = unknown).}

\item{informative}{Informative individuals selection can take 5 values:
\itemize{
\item 'AvAf' (available and affected),
\item 'AvOrAf' (available or affected),
\item 'Av' (available only),
\item 'Af' (affected only),
\item 'All' (all individuals)
\item A numeric/character vector of individuals id
\item A boolean
}}

\item{missid}{A character vector with the missing values identifiers.
All the id, dadid and momid corresponding to those values will be set
to \code{NA_character_}.}

\item{col_aff}{A string with the column name to use for the affection status.}

\item{reset}{If \code{TRUE}, the \code{id_inf} column is reset}
}
\value{
\subsection{When obj is a vector}{

A vector of individuals informative identifiers
}

\subsection{When obj is a Pedigree}{

A list containing the Pedigree object and the vector of individuals
identifiers.
The Pedigree object will have a new column named 'id_inf' containing 1 for
informative individuals and 0 otherwise.
}
}
\description{
Select the ids of the informative individuals.
}
\details{
Depending on the informative parameter, the function will extract
the ids of the informative individuals. In the case of a numeric vector,
the function will return the same vector. In the case of a boolean, the
function will return the ids of the individuals if TRUE, NA otherwise.
In the case of a string, the function will return the ids of the
corresponding informative individuals based on the avail and affected
columns.
}
\examples{
data("sampleped")
ped <- Pedigree(sampleped)
is_informative(ped, col_aff = "affection_aff")

}
