% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_informative.R
\docType{methods}
\name{is_informative}
\alias{is_informative}
\alias{is_informative,character-method}
\alias{is_informative,character}
\alias{is_informative,Pedigree-method}
\alias{is_informative,Pedigree}
\title{Is informative}
\usage{
is_informative(obj, ...)

\S4method{is_informative}{character}(obj, avail, affected, informative = "AvAf", missid = "0")

\S4method{is_informative}{Pedigree}(
  obj,
  col_aff = NULL,
  informative = "AvAf",
  missid = "0",
  reset = FALSE
)
}
\arguments{
\item{obj}{A pedigree object or a vector of subject identifiers.}

\item{...}{Additional arguments passed to methods}

\item{avail}{A numeric vector of availability status of each individual
(e.g., genotyped). The values are:
\itemize{
\item \code{0}  : unavailable
\item \code{1}  : available
\item \code{NA} : availability not known
}}

\item{affected}{A numeric vector of affection status of each individual
(e.g., genotyped). The values are:
\itemize{
\item \code{0}  : unaffected
\item \code{1}  : affected
\item \code{NA} : affection status not known
}}

\item{informative}{Informative individuals selection can take 5 values:
\itemize{
\item 'AvAf' (available and affected),
\item 'AvOrAf' (available or affected),
\item 'Av' (available only),
\item 'Af' (affected only),
\item 'All' (all individuals)
\item A numeric/character vector of individuals id
\item A boolean
}}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{"0"}.}

\item{col_aff}{A string with the column name to use for the affection status.}

\item{reset}{If \code{TRUE}, the \code{id_inf} column is reset}
}
\value{
\subsection{When obj is a vector}{

A vector of individuals informative identifiers
}

\subsection{When obj is a Pedigree}{

A list containing the Pedigree object and the vector of individuals
identifiers.
The Pedigree object will have a new column named 'id_inf' containing 1 for
informative individuals and 0 otherwise.
}
}
\description{
Select the ids of the informative individuals.
}
\details{
Depending on the informative parameter, the function will extract
the ids of the informative individuals. In the case of a numeric vector,
the function will return the same vector. In the case of a boolean, the
function will return the ids of the individuals if TRUE, NA otherwise.
In the case of a string, the function will return the ids of the
corresponding informative individuals based on the avail and affected
columns.
}
\examples{
data("sampleped")
ped <- pedigree(sampleped)
is_informative(ped, col_aff = "affected_aff")

}
