% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_informative.R
\docType{methods}
\name{is_informative}
\alias{is_informative}
\alias{is_informative,character-method}
\alias{is_informative,character}
\alias{is_informative,data.frame-method}
\alias{is_informative,data.frame}
\alias{is_informative,Pedigree-method}
\alias{is_informative,Pedigree}
\title{Is informative}
\usage{
is_informative(obj, ...)

\S4method{is_informative}{character}(obj, avail, affected, informative = "AvAf", missid = "0")

\S4method{is_informative}{data.frame}(obj, informative = "AvAf", missid = "0")

\S4method{is_informative}{Pedigree}(
  obj,
  column = NULL,
  informative = "AvAf",
  missid = "0",
  reset = FALSE
)
}
\arguments{
\item{obj}{A vector of individuals identifiers, a dataframe or a Pedigree
object.}

\item{...}{Other arguments passed to methods.}

\item{avail}{A vector of individuals availability (0, 1 or NA)}

\item{affected}{A vector of individuals affected status (0, 1 or NA)}

\item{informative}{Informative individuals selection can take 5 values:
'AvAf' (available and affected),
'AvOrAf' (available or affected),
'Av' (available only),
'Af' (affected only),
'All' (all individuals)
or a numeric/character vector of individuals id
or a boolean}

\item{missid}{A string with the missing id}

\item{column}{A string with the column name to use for the affected status.
This column must be in the scales fill.}

\item{reset}{Boolean defining if the inf column needs to be reset}
}
\value{
\subsection{When obj is a vector or a dataframe}{

A vector of individuals identifiers
}

\subsection{When obj is a Pedigree}{

A list containing the Pedigree object and the vector of individuals
identifiers.
The Pedigree object will have a new column named 'inf' containing 1 for
informative individuals and 0 otherwise.
}
}
\description{
Select the ids of the informative individuals.
}
\details{
Depending on the informative parameter, the function will extract
the ids of the informative individuals. In the case of a numeric vector,
the function will return the same vector. In the case of a boolean, the
function will return the ids of the individuals if TRUE, NA otherwise.
In the case of a string, the function will return the ids of the
corresponding informative individuals based on the avail and affected
columns.
}
\examples{
data("sampleped")
ped <- pedigree(sampleped)
is_informative(ped, column = "affected_aff")

}
