% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_parent}
\alias{is_parent}
\title{Check wich individuals are parents}
\usage{
is_parent(id, dadid, momid, missid = "0")
}
\arguments{
\item{id}{Vector of individual id}

\item{dadid}{Vector of father index}

\item{momid}{Vector of mother index}

\item{missid}{Missing value identifier}
}
\value{
A vector of boolean of the same size as \code{id}
with TRUE if the individual is a parent and FALSE otherwise
}
\description{
Check which individuals are parents.
}
