% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_parent}
\alias{is_parent}
\title{Check wich individuals are parents}
\usage{
is_parent(obj, ...)
}
\arguments{
\item{obj}{A vector of each subjects identifiers}
}
\value{
A vector of boolean of the same size as \code{id}
with TRUE if the individual is a parent and FALSE otherwise
}
\description{
Check which individuals are parents.
}
\keyword{internal}
