% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_parent}
\alias{is_parent}
\title{Check wich individuals are parents}
\usage{
is_parent(id, dadid, momid, missid = "0")
}
\arguments{
\item{id}{A vector of each subjects identifiers}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{"0"}.}
}
\value{
A vector of boolean of the same size as \code{id}
with TRUE if the individual is a parent and FALSE otherwise
}
\description{
Check which individuals are parents.
}
\keyword{internal}
