% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{is_valid}
\alias{is_valid}
\title{Pedigree validity method.}
\usage{
is_valid(object)
}
\description{
Check if the Pedigree object is valid.
}
\details{
It will check :
the fields of the slots
the values in the columns of the ped, rel and scale slot
$param object A Pedigree object.
$return A logical value or a character vector with the errors.
}
