% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindepth.R
\docType{methods}
\name{kindepth}
\alias{kindepth}
\alias{kindepth,character-method}
\alias{kindepth,character}
\alias{kindepth,Pedigree-method}
\alias{kindepth,Pedigree}
\title{Compute the depth of each subject in a pedigree}
\usage{
kindepth(obj, ...)

\S4method{kindepth}{character}(obj, dadid, momid, align_parents = FALSE)

\S4method{kindepth}{Pedigree}(obj, align_parents = FALSE)
}
\arguments{
\item{obj}{A pedigree object or a vector of subject identifiers.}

\item{...}{Additional arguments passed to methods}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{align_parents}{If \code{align_parents=T}, go one step further and try to
make both parents of each child have the same depth.
(This is not always possible).
It helps the drawing program by lining up pedigrees that 'join in the middle'
via a marriage.}
}
\value{
an integer vector containing the depth for each subject
}
\description{
Computes the depth of each subject in the pedigree.
}
\details{
Mark each person as to their depth in a pedigree; 0 for a founder, otherwise
\verb{depth = 1 + max(father's depth, mother's depth)}

In the case of an inbred pedigree a perfect alignment obeying
\code{extra=TRUE} may not exist.
}
\seealso{
\code{\link[=align]{align()}}
}
\author{
Terry Therneau
}
