% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindepth.R
\name{kindepth}
\alias{kindepth}
\title{Compute the depth of each subject in a pedigree}
\usage{
kindepth(id, dadid, momid, align = FALSE)
}
\arguments{
\item{id}{Identification code for each individual}

\item{dadid}{Id code for the father}

\item{momid}{Id code for the mother}

\item{align}{If `align=T`, go one step further and try to make both parents
of each child have the same depth.  (This is not always possible).
It helps the drawing program by lining up pedigrees that 'join in the middle'
via a marriage.}
}
\value{
an integer vector containing the depth for each subject
}
\description{
Computes the depth of each subject in the pedigree.
}
\details{
Mark each person as to their depth in a pedigree; 0 for a founder, otherwise
`depth = 1 + max(father's depth, mother's depth)`

In the case of an inbred pedigree a perfect alignment obeying
`extra=TRUE` may not exist.
}
\seealso{
`plot.pedigree`
}
\author{
Terry Therneau
}
\keyword{genetics}
