% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship.R
\docType{methods}
\name{kinship}
\alias{kinship}
\alias{kinship,character-method}
\alias{kinship,character}
\alias{kinship,Pedigree-method}
\alias{kinship,Pedigree}
\title{Compute a kinship matrix}
\usage{
kinship(obj, ...)

\S4method{kinship}{character}(obj, dadid, momid, sex, chrtype = "autosome")

\S4method{kinship}{Pedigree}(obj, chrtype = "autosome")
}
\arguments{
\item{obj}{A pedigree object or a vector of subject identifiers.}

\item{...}{Any number of optional arguments}

\item{dadid}{for each subject, the identifier of the biological father.
This is only used if \code{id} is a vector.}

\item{momid}{for each subject, the identifier of the biological mother.
This is only used if \code{id} is a vector.}

\item{sex}{vector of sex values coded as 1=male, 2=female}

\item{chrtype}{chromosome type.  The currently supported types are
'autosome' and 'X' or 'x'.}
}
\value{
\subsection{When obj is a vector}{

A matrix of kinship coefficients.
}

\subsection{When obj is a pedigree}{

A matrix of kinship coefficients ordered by families present
in the pedigree.
}
}
\description{
Compute the kinship matrix for a set of related autosomal subjects.  The
function is generic, and can accept a pedigree, pedigreeList, or vector as
the first argument.
}
\details{
The function will usually be called with a pedigree or pedigreeList; the
third form is provided for backwards compatability with an earlier release
of the library that was less capable.  The first argument is named \code{id}
for the same reason.  Note that when using the third form any information on
twins is not available to the function.

When called with a pedigree, the routine
will create a block-diagonal-symmetric sparse matrix object of class
\code{dsCMatrix}.  Since the \verb{[i, j]} value of the result is 0 for any two
unrelated individuals i and j and a \code{Matrix} utilizes sparse
representation, the resulting object is often orders of magnitude smaller
than an ordinary matrix.

Two genes G1 and G2 are identical by descent (IBD) if they are both physical
copies of the same ancestral gene; two genes are identical by state if they
represent the same allele.  So the brown eye gene that I inherited from my
mother is IBD with hers; the same gene in an unrelated individual is not.

The kinship coefficient between two subjects is the probability that a
randomly selected allele from a locus will be IBD between them. It is
obviously 0 between unrelated individuals. For an autosomal site and no
inbreeding it will be 0.5 for an individual with themselves, .25 between
mother and child, .125 between an uncle and neice, etc.

The computation is based on a recursive algorithm described in Lange, which
assumes that the founder alleles are all independent.
}
\section{References}{
 K Lange, Mathematical and Statistical Methods for
Genetic Analysis, Springer-Verlag, New York, 1997.
}

\examples{
data(sampleped)
ped <- pedigree(sampleped)
kinship(ped)

}
\seealso{
\code{\link[=make_famid]{make_famid()}}, \code{\link[=kindepth]{kindepth()}}
}
