% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_famid.R
\name{make_famid}
\alias{make_famid}
\title{Get family id}
\usage{
make_famid(obj, ...)
}
\arguments{
\item{id}{Identifier for each subject in the set of pedigrees}

\item{dadid}{Identifier for the father.  This will be 0 or '' for a
founder.}

\item{momid}{Identifer for the mother.}
}
\value{
An integer vector giving family groupings
}
\description{
Construct a family id from pedigree information
}
\details{
Create a vector of length n, giving the family 'tree' number of each
subject.  If the pedigree is totally connected, then everyone will end up in
tree 1, otherwise the tree numbers represent the disconnected subfamilies.
Singleton subjects give a zero for family number.

This command is depricated.  The kinship command now can be applied directly
to pedigreeList objects.
}
\seealso{
\code{kinship}
}
\author{
Terry Therneau
}
\keyword{genetics}
