% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_kin_inf.R
\name{max_kin_inf}
\alias{max_kin_inf}
\title{Kinship computation}
\usage{
max_kin_inf(obj, ...)
}
\arguments{
\item{...}{Other parameters passed to \code{\link{is_informative}}
when using a pedigree object}

\item{id}{A vector of individuals id}

\item{dadid}{A vector of individuals dadid}

\item{momid}{A vector of individuals momid}

\item{sex}{A vector of the individuals sex}

\item{avail}{A vector of individuals availability (0, 1 or NA)}

\item{affected}{A vector of individuals affected status (0, 1 or NA)}

\item{informative}{Informative individuals selection can take 3 values:
'AvAf' (available and affected),
'AvOrAf' (available or affected),
'Av' (available only),
'Af' (affected only),
'All' (all individuals)
or a numeric vector of individuals id
or a boolean}
}
\value{
Dataframe with selected individuals
}
\description{
Compute the kinship between the informative individuals and
all the others.
}
\examples{
data(sampleped)
df <- max_kin_inf(sampleped, informative = 'Av/Af')
summary(df$kin)

}
