% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_kin_inf.R
\name{max_kin_inf}
\alias{max_kin_inf}
\title{Kinship computation}
\usage{
max_kin_inf(df, informative = "AvAf")
}
\arguments{
\item{df}{Dataframe with individuals to select}

\item{informative}{Informative individuals selection can take 3 values:
'AvAf' (available and affected),
'AvOrAf' (available or affected),
'Av' (available only),
'Af' (affected only),
'All' (all individuals)
or a numeric vector of individuals id
or a boolean}
}
\value{
Dataframe with selected individuals
}
\description{
Compute the kinship between the informative individuals and
all the others.
}
\examples{
data(sample.ped)
df <- max_kin_inf(sample.ped, informative = 'Av/Af')
summary(df$kin)

}
