% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_inf.R
\docType{methods}
\name{min_dist_inf}
\alias{min_dist_inf}
\alias{min_dist_inf,character-method}
\alias{min_dist_inf,character}
\alias{min_dist_inf,Pedigree-method}
\alias{min_dist_inf,Pedigree}
\title{Minimum distance to the informative individuals}
\usage{
min_dist_inf(obj, ...)

\S4method{min_dist_inf}{character}(obj, dadid, momid, sex, avail, affected, informative = "AvAf")

\S4method{min_dist_inf}{Pedigree}(
  obj,
  col_aff = NULL,
  informative = "AvAf",
  missid = "0",
  reset = FALSE,
  ...
)
}
\arguments{
\item{obj}{A pedigree object or a vector of subject identifiers.}

\item{...}{Additional arguments passed to methods}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{sex}{A character, factor or numeric vector corresponding to
the gender of the individuals. The following values are recognized:
\itemize{
\item character() or factor() : "f", "m", "woman", "man", "male", "female",
"unknown", "terminated"
\item numeric() : 1 = "male", 2 = "female", 3 = "unknown", 4 = "terminated"
}}

\item{avail}{A numeric vector of availability status of each individual
(e.g., genotyped). The values are:
\itemize{
\item \code{0}  : unavailable
\item \code{1}  : available
\item \code{NA} : availability not known
}}

\item{affected}{A numeric vector of affection status of each individual
(e.g., genotyped). The values are:
\itemize{
\item \code{0}  : unaffected
\item \code{1}  : affected
\item \code{NA} : affection status not known
}}

\item{informative}{Informative individuals selection can take 5 values:
\itemize{
\item 'AvAf' (available and affected),
\item 'AvOrAf' (available or affected),
\item 'Av' (available only),
\item 'Af' (affected only),
\item 'All' (all individuals)
\item A numeric/character vector of individuals id
\item A boolean
}}

\item{col_aff}{A string with the column name to use for the affection status.}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree (i.e. \code{dadid} and \code{momid} equal to
the value of this variable).  The default for \code{missid} is \code{"0"}.}

\item{reset}{If TRUE, the \code{kin} and if \code{id_inf} columns is reset}
}
\value{
\subsection{When obj is a vector}{

A vector of the minimum distance between the informative individuals
and all the others corresponding to the order of the individuals in the
\code{obj} vector.
}

\subsection{When obj is a Pedigree}{

The Pedigree object with a new column named 'kin' containing the kinship
degree.
}
}
\description{
Compute the minimum distance between the informative
individuals and all the others.
This distance is a transformation of the maximum kinship degree between
the informative individuals and all the others.
This transformation is done by taking the log2 of the inverse of the
maximum kinship degree.

\eqn{minDist = log2(1 / \max(kinship))}

Therefore, the minimum distance is 0 when the maximum kinship is 1 and
is infinite when the maximum kinship is 0. For siblings, the kinship value
is 0.5 and the minimum distance is 1. Each time the kinship degree is divided
by 2, the minimum distance is increased by 1.
}
\examples{
data(sampleped)
ped <- pedigree(sampleped)
min_dist_inf(ped, col_aff = "affected_aff")$ped

}
\seealso{
\code{\link[=kinship]{kinship()}}
}
