% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_inf.R
\name{min_dist_inf}
\alias{min_dist_inf}
\alias{min_dist_inf,character-method}
\alias{min_dist_inf,Pedigree-method}
\alias{min_dist_inf,Ped-method}
\title{Minimum distance to the informative individuals}
\usage{
\S4method{min_dist_inf}{character}(obj, dadid, momid, sex, id_inf)

\S4method{min_dist_inf}{Pedigree}(obj, col_aff = NULL, informative = "AvAf", reset = FALSE, ...)

\S4method{min_dist_inf}{Ped}(obj, informative = "AvAf", reset = FALSE)
}
\arguments{
\item{obj}{A character vector with the id of the individuals or a
\code{data.frame} with all the informations in corresponding columns.}

\item{...}{Additional arguments}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{A vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{sex}{A character, factor or numeric vector corresponding to
the gender of the individuals. This will be transformed to an ordered factor
with the following levels: \code{male} < \code{female} < \code{unknown} < `terminated
The following values are recognized:
\itemize{
\item character() or factor() : "f", "m", "woman", "man", "male", "female",
"unknown", "terminated"
\item numeric() : 1 = "male", 2 = "female", 3 = "unknown", 4 = "terminated"
}}

\item{id_inf}{An identifiers vector of informative individuals.}

\item{col_aff}{A character vector with the name of the column to be used
for the affection status.}

\item{informative}{Informative individuals selection can take 5 values:
\itemize{
\item 'AvAf' (available and affected),
\item 'AvOrAf' (available or affected),
\item 'Av' (available only),
\item 'Af' (affected only),
\item 'All' (all individuals)
\item A numeric/character vector of individuals id
\item A boolean
}}

\item{reset}{If TRUE, the \code{kin} and if \code{isinf} columns is reset}
}
\value{
\subsection{When obj is a vector}{

A vector of the minimum distance between the informative individuals
and all the others corresponding to the order of the individuals in the
\code{obj} vector.
}

\subsection{When obj is a Pedigree}{

The Pedigree object with a new slot named 'kin' containing the minimum
distance between each individuals and the informative individuals.
The \code{isinf} slot is also updated with the informative individuals.
}
}
\description{
Compute the minimum distance between the informative
individuals and all the others.
This distance is a transformation of the maximum kinship degree between
the informative individuals and all the others.
This transformation is done by taking the log2 of the inverse of the
maximum kinship degree.

\eqn{minDist = log2(1 / \max(kinship))}

Therefore, the minimum distance is 0 when the maximum kinship is 1 and
is infinite when the maximum kinship is 0. For siblings, the kinship value
is 0.5 and the minimum distance is 1. Each time the kinship degree is divided
by 2, the minimum distance is increased by 1.
}
\examples{

min_dist_inf(
     c("A", "B", "C", "D", "E"),
     c("C", "D", "0", "0", "0"),
     c("E", "E", "0", "0", "0"),
     sex = c(1, 2, 1, 2, 1),
     id_inf = c("D", "E")
)

data(sampleped)
ped <- Pedigree(sampleped)
kin(ped(min_dist_inf(ped, col_aff = "affection_mods")))
}
\seealso{
\code{\link[=kinship]{kinship()}}
}
