% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\name{norm_ped}
\alias{norm_ped}
\title{Normalise dataframe}
\usage{
norm_ped(ped_df, na_strings = c("NA", ""), missid = "0", try_num = FALSE)
}
\arguments{
\item{ped_df}{A data.frame with the individuals informations.
The minimum columns required are \code{id}, \code{dadid}, \code{momid} and
\code{sex}.
The \code{family} column can also be used to specify the family of the
individuals and will be merge to the \code{id} field separated by an
underscore.
The following columns are also recognize \code{steril}, \code{avail},
\code{status}, \code{affected}.
They respectively correspond to the sterilisation status,
the availability status, the death status and the affection status
of the individuals. The values recognized for those columns are \code{1} or
\code{0}.}

\item{na_strings}{Vector of strings to be considered as NA values}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{"0"}.}

\item{try_num}{Boolean defining if the function should try to convert
all the columns to numeric.}
}
\value{
A dataframe with the errors identified in the \code{error} column
}
\description{
Normalise dataframe for pedigree object
}
\details{
Normalise a dataframe and check for columns correspondance
to be able to use it as an input to create pedigree object.
Multiple test are done and errors are checked.
Sex is calculated based in the \code{gender} column the following notations
are accepted: f, woman, female, 2 and m, man, male, 1.
The \code{steril} column need to be a boolean either TRUE, FALSE or 'NA'.
Will be considered available any individual with no 'NA' values in the
\code{available} column.
Duplicated \code{indId} will nullify the relationship of the individual.
All individuals with errors will be remove from the dataframe and will
be transfered to the error dataframe.
}
