% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_child.R
\docType{methods}
\name{num_child}
\alias{num_child}
\alias{num_child,character-method}
\alias{num_child,character}
\alias{num_child,data.frame-method}
\alias{num_child,data.frame}
\alias{num_child,Pedigree-method}
\alias{num_child,Pedigree}
\title{Number of child}
\usage{
num_child(obj, ...)

\S4method{num_child}{character}(obj, dadid, momid, relation = NULL, missid = "0")

\S4method{num_child}{data.frame}(obj, relation = NULL, ...)

\S4method{num_child}{Pedigree}(obj, reset = FALSE)
}
\arguments{
\item{obj}{A pedigree object, a dataframe or a vector of the individuals}

\item{...}{Additional arguments to be passed to methods.}

\item{dadid}{The id of the father}

\item{momid}{The id of the mother}

\item{relation}{A matrix with 3 required columns (id1, id2, code) specifying
special relationship between pairs of individuals.
Codes: 1=Monozygotic twin, 2=Dizygotic twin, 3=twin of unknown zygosity,
4=Spouse.}

\item{missid}{Missing id code}

\item{reset}{If TRUE, the num_child columns are reset}
}
\value{
\subsection{When obj is a vector or a dataframe}{

A dataframe with the columns \code{num_child_dir}, \code{num_child_ind} and
\code{num_child_tot} giving respectively the direct, indirect and total number
of child.
}

\subsection{When obj is a pedigree object}{

An updated pedigree object with the columns \code{num_child_dir}, \code{num_child_ind}
and \code{num_child_tot} added to the pedigree.
}
}
\description{
Compute the number of child per individual
}
\details{
Compute the number of direct child but also the number
of indirect child given by the ones related with the linked spouses.
If a relation ship matrix is given, then even if no children is present
between 2 spouses, the indirect childs will still be added.
}
