% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_child.R
\docType{methods}
\name{num_child}
\alias{num_child}
\alias{num_child,character-method}
\alias{num_child,character}
\alias{num_child,Pedigree-method}
\alias{num_child,Pedigree}
\title{Number of child}
\usage{
num_child(obj, ...)

\S4method{num_child}{character}(obj, dadid, momid, rel_df = NULL, missid = "0")

\S4method{num_child}{Pedigree}(obj, reset = FALSE)
}
\arguments{
\item{obj}{A pedigree object or a vector of subject identifiers.}

\item{...}{Additional arguments passed to methods}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{rel_df}{A data.frame with the special relationships between
individuals.
The minimum columns required are \code{id1}, \code{id2} and \code{code}. The \code{family}
column can also be used to specify the family of the individuals.
The code values are:
\itemize{
\item \code{1} = Monozygotic twin
\item \code{2} = Dizygotic twin
\item \code{3} = twin of unknown zygosity
\item \code{4} = Spouse
}

The value relation code recognized by the function are the one defined
by the \code{\link[=rel_code_to_factor]{rel_code_to_factor()}} function.}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree (i.e. \code{dadid} and \code{momid} equal to
the value of this variable).  The default for \code{missid} is \code{"0"}.}

\item{reset}{If TRUE, the \code{num_child_tot}, \code{num_child_ind} and
the \code{num_child_dir} columns are reset.}
}
\value{
\subsection{When obj is a vector}{

A dataframe with the columns \code{num_child_dir}, \code{num_child_ind} and
\code{num_child_tot} giving respectively the direct, indirect and total number
of child.
}

\subsection{When obj is a pedigree object}{

An updated pedigree object with the columns \code{num_child_dir}, \code{num_child_ind}
and \code{num_child_tot} added to the pedigree \code{ped} slot.
}
}
\description{
Compute the number of child per individual
}
\details{
Compute the number of direct child but also the number
of indirect child given by the ones related with the linked spouses.
If a relation ship matrix is given, then even if no children is present
between 2 spouses, the indirect childs will still be added.
}
