% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllValidity.R
\name{paste0max}
\alias{paste0max}
\title{Print0 to max}
\usage{
paste0max(x, max = 5, sep = "", ...)
}
\arguments{
\item{x}{A vector.}

\item{max}{The maximum number of elements to print.}

\item{...}{Additional arguments passed to print0}
}
\value{
The character vector aggregated until the maximum is reached.
}
\description{
Print0 the elements inside a vector until a maximum is reached.
}
\keyword{internal}
