% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructor.R
\name{Ped}
\alias{Ped}
\title{Constructor for the Ped class}
\usage{
Ped(
  id,
  sex,
  dadid,
  momid,
  family = NA,
  steril = NA,
  status = NA,
  avail = NA,
  affected = NA,
  useful = NA,
  kin = NA,
  missid = NA
)
}
\arguments{
\item{id}{A character vector with the id of the individuals.}

\item{sex}{A factor vector with the sex of the individuals (i.e. \code{male},
\code{female}, \code{unknown} or \code{terminated}).}

\item{dadid}{A character vector with the id of the father of the individuals.}

\item{momid}{A character vector with the id of the mother of the individuals.}

\item{family}{A character vector with the family of the individuals.}

\item{steril}{A numeric vector with the sterilisation status of the
individuals (i.e. \code{0} = not sterilised, \code{1} = sterilised, \code{NA} = unknown).}

\item{status}{A numeric vector with the affection status of the
individuals (i.e. \code{0} = alive, \code{1} = dead, \code{NA} = unknown).}

\item{avail}{A numeric vector with the availability status of the
individuals (i.e. \code{0} = not available, \code{1} = available, \code{NA} = unknown).}
}
\value{
A Ped object.
}
\description{
Constructor for the Ped class
}
\seealso{
\code{\link[=Pedigree]{Pedigree()}}
}
