% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/AllConstructor.R,
%   R/AllAccessors.R
\docType{class}
\name{Ped-class}
\alias{Ped-class}
\alias{Ped}
\alias{Ped,data.frame-method}
\alias{Ped,data.frame}
\alias{Ped,character_OR_integer-method}
\alias{Ped,character_OR_integer}
\alias{Ped,missing-method}
\alias{Ped,missing}
\alias{mcols<-,Ped,list-method}
\alias{mcols<-,Ped,data.frame-method}
\alias{famid}
\alias{famid,Ped-method}
\alias{id}
\alias{id,Ped-method}
\alias{dadid}
\alias{dadid,Ped-method}
\alias{momid}
\alias{momid,Ped-method}
\alias{sex}
\alias{sex,Ped-method}
\alias{affected}
\alias{affected,Ped-method}
\alias{avail}
\alias{avail,Ped-method}
\alias{kin}
\alias{kin,Ped-method}
\alias{isinf}
\alias{isinf,Ped-method}
\alias{useful}
\alias{useful,Ped-method}
\title{S4 class to represent the identity informations in a Pedigree.}
\usage{
Ped(obj, ...)

\S4method{Ped}{data.frame}(obj, cols_used_init = FALSE, cols_used_del = FALSE)

\S4method{Ped}{character_OR_integer}(
  obj,
  sex,
  dadid,
  momid,
  famid = NA,
  steril = NA,
  status = NA,
  avail = NA,
  affected = NA,
  missid = NA_character_
)

\S4method{Ped}{missing}(obj)

\S4method{mcols}{Ped,list}(x) <- value

\S4method{mcols}{Ped,data.frame}(x) <- value

famid(x)

id(x)

dadid(x)

momid(x)

sex(x)

affected(x)

avail(x)

kin(x)

isinf(x)

useful(x)
}
\arguments{
\item{obj}{A character vector with the id of the individuals or a
\code{data.frame} with all the informations in corresponding columns.}

\item{cols_used_init}{Boolean defining if the columns that will be used
should be initialised to NA.}

\item{cols_used_del}{Boolean defining if the columns that will be used
should be deleted.}

\item{sex}{A character, factor or numeric vector corresponding to
the gender of the individuals. This will be transformed to an ordered factor
with the following levels: \code{male} < \code{female} < \code{unknown} < `terminated
The following values are recognized:
\itemize{
\item character() or factor() : "f", "m", "woman", "man", "male", "female",
"unknown", "terminated"
\item numeric() : 1 = "male", 2 = "female", 3 = "unknown", 4 = "terminated"
}}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{famid}{A character vector with the family identifiers of the
individuals.}

\item{steril}{A logical vector with the sterilisation status of the
individuals
(i.e. \code{FALSE} = not sterilised, \code{TRUE} = sterilised, \code{NA} = unknown).}

\item{status}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = alive, \code{TRUE} = dead, \code{NA} = unknown).}

\item{avail}{A logical vector with the availability status of the
individuals
(i.e. \code{FALSE} = not available, \code{TRUE} = available, \code{NA} = unknown).}

\item{affected}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = unaffected, \code{TRUE} = affected, \code{NA} = unknown).}

\item{missid}{A character vector with the missing values identifiers.
All the id, dadid and momid corresponding to those values will be set
to \code{NA_character_}.}

\item{x}{A Ped object.}

\item{value}{A data.frame with the metadata.}
}
\value{
A Ped object.

A Ped object.

A Ped object with the metadata set.

A Ped object with the metadata set.

A character vector with the famid of each individual.

A character vector with the id of each individual.

A character vector with the dadid of each individual.

A character vector with the momid of each individual.

A character vector with the sex of each individual.

A numeric vector with the affected of each individual.

A numeric vector with the avail of each individual.

A numeric vector with the minimum kinship distance
value of each individual towards the informative
individuals.

A numeric vector with the saying if the individual
is informative or not.

A numeric vector with the saying if the individual
is informative or not.
}
\description{
A Ped object is a list of identity informations
of the individuals in the pedigree.
It is used to create a Pedigree object.
The minimal needed informations are \code{id}, \code{dadid}, \code{momid} and \code{sex}.
If a \code{famid} is provided, the individuals \code{id} will be aggregated
to the \code{famid} character to ensure the uniqueness of the \code{id}.
The other slots are used to store recognized informations.
Additional columns can be added to the Ped object and will be
stored in the \code{elementMetadata} slot of the Ped object.

Constructor for the Ped class
You either need to provide a vector of the same size for each slot
or a \code{data.frame} with the corresponding columns.
The metadata will correspond to the columns that do not correspond
to the Ped slots.
}
\section{Slots}{

\describe{
\item{\code{id}}{A character vector with the id of the individuals.}

\item{\code{dadid}}{A character vector with the id of the father of the individuals.}

\item{\code{momid}}{A character vector with the id of the mother of the individuals.}

\item{\code{sex}}{An ordered factor vector for the sex of the individuals
(i.e. \code{male} < \code{female} < \code{unknown} < \code{terminated}).}

\item{\code{famid}}{A character vector with the family identifiers of the
individuals (optional).}

\item{\code{steril}}{A logical vector with the sterilisation status of the
individuals
(i.e. \code{FALSE} = not sterilised, \code{TRUE} = sterilised, \code{NA} = unknown).}

\item{\code{status}}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = alive, \code{TRUE} = dead, \code{NA} = unknown).}

\item{\code{avail}}{A logical vector with the availability status of the
individuals
(i.e. \code{FALSE} = not available, \code{TRUE} = available, \code{NA} = unknown).}

\item{\code{affected}}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = not affected, \code{TRUE} = affected, \code{NA} = unknown).}

\item{\code{useful}}{A logical vector with the usefulness status of the
individuals
(i.e. \code{FALSE} = not useful, \code{TRUE} = useful).}

\item{\code{kin}}{A numeric vector with minimal kinship value between the
individuals and the useful individuals.}

\item{\code{num_child_tot}}{A numeric vector with the total number of children
of the individuals.}

\item{\code{num_child_dir}}{A numeric vector with the number of children
of the individuals.}

\item{\code{num_child_ind}}{A numeric vector with the number of children
of the individuals.}
}}

\examples{
data(sampleped)
Ped(sampleped)
Ped(
    obj = c("1", "2", "3", "4", "5", "6"),
    dadid = c("4", "4", "6", "0", "0", "0"),
    momid = c("5", "5", "5", "0", "0", "0"),
    sex = c(1, 2, 3, 1, 2, 1),
    missid = "0"
)
Ped()
}
\seealso{
\code{\link[=Pedigree]{Pedigree()}}

\code{\link[=Pedigree]{Pedigree()}}
}
