% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R, R/AllAccessors.R,
%   R/AllConstructor.R
\docType{class}
\name{Ped-class}
\alias{Ped-class}
\alias{mcols<-,Ped,list-method}
\alias{mcols<-,Ped,data.frame-method}
\alias{Ped}
\alias{Ped,data.frame-method}
\alias{Ped,data.frame}
\alias{Ped,character-method}
\alias{Ped,character}
\title{S4 class to represent the identity informations in a Pedigree.}
\usage{
\S4method{mcols}{Ped,list}(x) <- value

\S4method{mcols}{Ped,data.frame}(x) <- value

Ped(obj, ...)

\S4method{Ped}{data.frame}(obj, cols_used_init = FALSE, cols_used_del = FALSE)

\S4method{Ped}{character}(
  obj,
  sex,
  dadid,
  momid,
  family = NA,
  steril = NA,
  status = NA,
  avail = NA,
  affected = NA
)
}
\arguments{
\item{x}{A Ped object.}

\item{value}{A data.frame with the metadata.}

\item{obj}{A character vector with the id of the individuals or a
\code{data.frame} with all the informations in corresponding columns.}

\item{sex}{A factor vector with the sex of the individuals (i.e. \code{male},
\code{female}, \code{unknown} or \code{terminated}).}

\item{dadid}{A character vector with the id of the father of the individuals.}

\item{momid}{A character vector with the id of the mother of the individuals.}

\item{family}{A character vector with the family of the individuals.}

\item{steril}{A numeric vector with the sterilisation status of the
individuals (i.e. \code{0} = not sterilised, \code{1} = sterilised, \code{NA} = unknown).}

\item{status}{A numeric vector with the affection status of the
individuals (i.e. \code{0} = alive, \code{1} = dead, \code{NA} = unknown).}

\item{avail}{A numeric vector with the availability status of the
individuals (i.e. \code{0} = not available, \code{1} = available, \code{NA} = unknown).}
}
\value{
A Ped object.

A Ped object with the metadata set.

A Ped object with the metadata set.

A Ped object.
}
\description{
A Ped object is a list of identity informations
of the individuals in the pedigree.
It is used to create a Pedigree object.
The minimal needed informations are \code{id}, \code{dadid}, \code{momid} and \code{sex}.
If a \code{family} is provided, the individuals \code{id} will be aggregated
to the \code{family} character to ensure the uniqueness of the \code{id}.
The other slots are used to store recognized informations.
Additional columns can be added to the Ped object and will be
stored in the \code{meta} slot of the Ped object.

Constructor for the Ped class
If a \code{data.frame} is provided, the metadata will correspond to the columns
that do not correspond to the Ped slots.
}
\section{Slots}{

\describe{
\item{\code{id}}{A character vector with the id of the individuals.}

\item{\code{dadid}}{A character vector with the id of the father of the individuals.}

\item{\code{momid}}{A character vector with the id of the mother of the individuals.}

\item{\code{sex}}{A factor vector for the sex of the individuals (i.e. \code{male},
\code{female}, \code{unknown} or \code{terminated}).}

\item{\code{family}}{A character vector with the family of the individuals.}

\item{\code{steril}}{A numeric vector with the sterilisation status of the
individuals (i.e. \code{0} = not sterilised, \code{1} = sterilised, \code{NA} = unknown).}

\item{\code{status}}{A numeric vector with the affection status of the
individuals (i.e. \code{0} = alive, \code{1} = dead, \code{NA} = unknown).}

\item{\code{avail}}{A numeric vector with the availability status of the
individuals (i.e. \code{0} = not available, \code{1} = available, \code{NA} = unknown).}

\item{\code{affected}}{A numeric vector with the affection status of the
individuals (i.e. \code{0} = not affected, \code{1} = affected, \code{NA} = unknown).}

\item{\code{useful}}{A numeric vector with the usefulness status of the
individuals (i.e. \code{0} = not useful, \code{1} = useful).}

\item{\code{kin}}{A numeric vector with minimal kinship value between the
individuals and the useful individuals.}

\item{\code{num_child_total}}{A numeric vector with the total number of children
of the individuals.}

\item{\code{num_child_direct}}{A numeric vector with the number of children
of the individuals.}

\item{\code{num_child_indirect}}{A numeric vector with the number of children
of the individuals.}
}}

\seealso{
\code{\link[=Pedigree]{Pedigree()}}

\code{\link[=Pedigree]{Pedigree()}}
}
