% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_to_plotdf.R
\name{ped_to_plotdf}
\alias{ped_to_plotdf}
\title{Convert a Pedigree to a data frame of element to plot}
\usage{
ped_to_plotdf(obj, ...)
}
\arguments{
\item{obj}{A Pedigree object}

\item{...}{Other arguments to pass to the function}

\item{pconnect}{when connecting parent to children the program will try to
make the connecting line as close to vertical as possible, subject to it
lying inside the endpoints of the line that connects the children by at
least \code{pconnect} people.  Setting this option to a large number will
force the line to connect at the midpoint of the children.}

\item{branch}{defines how much angle is used to connect various levels of
nuclear families.}

\item{aff_mark}{if TRUE, add a aff_mark to each box corresponding to the
value of the affection column for each filling scale.}

\item{label}{if not NULL, add a label to each box corresponding to the
value of the column given.}
}
\value{
A list containing the data frame and the user coordinates.
}
\description{
Convert a Pedigree to a data frame of element to plot
}
\examples{
data(sampleped)
ped1 <- Pedigree(sampleped[sampleped$famid == 1,])
ped_to_plotdf(ped1)
}
\seealso{
\code{\link[=plot_fromdf]{plot_fromdf()}}
\code{\link[=ped_to_legdf]{ped_to_legdf()}}
}
