% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeClass.R
\docType{class}
\name{Pedigree-class}
\alias{Pedigree-class}
\title{S4 class to represent a pedigree.}
\value{
A Pedigree object.
}
\description{
A pedigree is a ensemble of individuals linked to each other into
a family tree.
}
\details{
They are created from a data.frame containing the individuals informations
and a relation ship data.frame for the special links between individuals.
A list of scales can be provided to create a legend.
To create a Pedigree object, use the function \code{\link[=pedigree]{pedigree()}}.
}
\section{Slots}{

\describe{
\item{\code{ped}}{A data.frame with the individuals informations. The minimum
columns required are 'id', 'dadid', 'momid' and 'sex'. Other columns can be
added to the data.frame and will be recognised by the functions. Some
errors can be detected by the validity function and some of them can be
corrected and others will be added to a dedicated column.}

\item{\code{rel}}{A data.frame for the special relationship between individuals.
The minimum columns required are 'id1', 'id2' and 'code'.}

\item{\code{scales}}{A data.frame to use for the affection status. This data.frame
is generated by the function \code{\link[=generate_aff_inds]{generate_aff_inds()}} followed by
\code{\link[=generate_colors]{generate_colors()}}.}

\item{\code{hints}}{List of two elements.
\itemize{
\item \strong{order} is a numeric vector with one element per subject in the
pedigree.  It determines the relative order of subjects within a sibship, as
well as the relative order of processing for the founder couples. (For this
latter, the female founders are ordered as though they were sisters).
\item \strong{spouse} is a matrix with one row per hinted marriage, usually only
a few marriages in a pedigree will need an added hint, for instance reverse
the plot order of a husband/wife pair. Each row contains the index of the
left spouse, the right hand spouse, and the anchor
(i.e : \code{1} = left, \code{2} = right, \code{0} = either).
}}
}}

\seealso{
\code{\link[=pedigree]{pedigree()}}
}
