% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllConstructor.R, R/AllAccessors.R
\docType{methods}
\name{Pedigree}
\alias{Pedigree}
\alias{Pedigree,character_OR_integer-method}
\alias{Pedigree,character}
\alias{Pedigree,data.frame-method}
\alias{Pedigree,data.frame}
\alias{Pedigree,missing-method}
\alias{Pedigree,missing}
\alias{famid,Pedigree-method}
\alias{horder}
\alias{horder,Pedigree-method}
\title{Create a Pedigree object}
\usage{
Pedigree(obj, ...)

\S4method{Pedigree}{character_OR_integer}(
  obj,
  dadid,
  momid,
  sex,
  famid = NA,
  avail = NULL,
  affected = NULL,
  status = NULL,
  steril = NULL,
  rel_df = NULL,
  missid = NA_character_,
  col_aff = "affection",
  normalize = TRUE,
  ...
)

\S4method{Pedigree}{data.frame}(
  obj = data.frame(indId = character(), fatherId = character(), motherId = character(),
    gender = numeric(), family = character(), available = numeric(), vitalStatus =
    numeric(), affection = numeric(), sterilisation = numeric()),
  rel_df = data.frame(id1 = character(), id2 = character(), code = numeric(), famid =
    character()),
  cols_ren_ped = list(indId = "id", fatherId = "dadid", motherId = "momid", family =
    "famid", gender = "sex", sterilisation = "steril", affection = "affected", available
    = "avail", vitalStatus = "status"),
  cols_ren_rel = list(id1 = "indId1", id2 = "indId2", famid = "family"),
  hints = list(horder = NULL, spouse = NULL),
  normalize = TRUE,
  missid = NA_character_,
  col_aff = "affection",
  ...
)

\S4method{Pedigree}{missing}(obj)

\S4method{famid}{Pedigree}(x)

horder(object)
}
\arguments{
\item{obj}{A vector of the individuals identifiers or a data.frame
with the individuals informations. See \code{\link[=Ped]{Ped()}} for more informations.}

\item{...}{Other arguments to pass to the function \code{generate_colors}.}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{A vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{sex}{A character, factor or numeric vector corresponding to
the gender of the individuals. This will be transformed to an ordered factor
with the following levels: \code{male} < \code{female} < \code{unknown} < `terminated
The following values are recognized:
\itemize{
\item character() or factor() : "f", "m", "woman", "man", "male", "female",
"unknown", "terminated"
\item numeric() : 1 = "male", 2 = "female", 3 = "unknown", 4 = "terminated"
}}

\item{famid}{A character vector with the family identifiers of the
individuals. If provide, will be aggregated to the individuals
identifiers separated by an underscore.}

\item{avail}{A logical vector with the availability status of the
individuals
(i.e. \code{FALSE} = not available, \code{TRUE} = available, \code{NA} = unknown).}

\item{affected}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = unaffected, \code{TRUE} = affected, \code{NA} = unknown).
Can also be a data.frame with the same length as \code{obj}. If it is a
matrix, it will be converted to a data.frame and the columns will be
named after the \code{col_aff} argument.}

\item{status}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = alive, \code{TRUE} = dead, \code{NA} = unknown).}

\item{steril}{A logical vector with the sterilisation status of the
individuals
(i.e. \code{FALSE} = not sterilised, \code{TRUE} = sterilised, \code{NA} = unknown).}

\item{rel_df}{A data.frame with the special relationships between
individuals. See \code{\link[=Rel]{Rel()}} for more informations.
The minimum columns required are \code{id1}, \code{id2} and \code{code}.
The \code{famid} column can also be used to specify the family
of the individuals.
If a matrix is given, the columns needs to be ordered as
\code{id1}, \code{id2}, \code{code} and \code{famid}.
The code values are:
\itemize{
\item \code{1} = Monozygotic twin
\item \code{2} = Dizygotic twin
\item \code{3} = twin of unknown zygosity
\item \code{4} = Spouse
}

The value relation code recognized by the function are the one defined
by the \code{\link[=rel_code_to_factor]{rel_code_to_factor()}} function.}

\item{missid}{A character vector with the missing values identifiers.
All the id, dadid and momid corresponding to those values will be set
to \code{NA_character_}.}

\item{col_aff}{A character vector with the name of the column to be used
for the affection status.
If \code{affected} is a data.frame, it will be overwritten by the column
names of the data.frame.}

\item{normalize}{A logical to know if the data should be normalised.}

\item{cols_ren_ped}{A named list with the columns to rename for the
pedigree dataframe. This is useful if you want to use a dataframe with
different column names. The names of the list should be the new column
names and the values should be the old column names. The default values
are to be used with \code{normalize = TRUE}.}

\item{cols_ren_rel}{A named list with the columns to rename for the
relationship matrix. This is useful if you want to use a dataframe with
different column names. The names of the list should be the new column
names and the values should be the old column names.}

\item{hints}{A Hints object or a named list containing \code{horder} and
\code{spouse}.}

\item{object}{A Pedigree object.}
}
\value{
A Pedigree object.

The slot \code{horder} present in the \code{Hints} slot of
a Pedigree object.
}
\description{
Main constructor of the package.
This constructor help to create a \code{Pedigree} object from
different \code{data.frame} or a set of vectors.

If any errors are found in the data, the function will return
the data.frame with the errors of the Ped object and the
Rel object.

If the normalization is set to \code{TRUE}, then the data will be
standardized using the function \code{norm_ped()} and \code{norm_rel()}.

If a data.frame is given, the columns names needed will depend if
the normalization is selected or not. If the normalization is selected,
the columns names needed are as follow and if not the columns names
needed are in parenthesis:
\itemize{
\item \code{indID}: the individual identifier (\code{id})
\item \code{fatherId}: the identifier of the biological father (\code{dadid})
\item \code{motherId}: the identifier of the biological mother (\code{momid})
\item \code{gender}: the sex of the individual (\code{sex})
\item \code{family}: the family identifier of the individual (\code{famid})
\item \code{sterilisation}: the sterilisation status of the individual (\code{steril})
\item \code{available}: the availability status of the individual (\code{avail})
\item \code{vitalStatus}: the death status of the individual (\code{status})
\item \code{affection}: the affection status of the individual (\code{affected})
\item \code{...}: other columns that will be stored in the \code{elementMetadata} slot
}

The minimum columns required are :
\itemize{
\item \code{indID} / \code{id}
\item \code{fatherId} / \code{dadid}
\item \code{motherId} / \code{momid}
\item \code{gender} / \code{sex}
The \code{family} / \code{famid} column can also be used to specify the family of the
individuals and will be merge to the \code{indId} / \code{id} field separated by an
underscore.
The columns \code{sterilisation}, \code{available}, \code{vitalStatus}, \code{affection}
will be transformed with the \code{\link[=vect_to_binary]{vect_to_binary()}} function when the
normalisation is selected.
If you do not use the normalisation, the columns will be checked to
be \code{0} or \code{1}.
}

Pedigree horder accessors
}
\examples{
Pedigree(
   obj = c("1", "2", "3", "4", "5", "6"),
   dadid = c("4", "4", "6", "0", "0", "0"),
   momid = c("5", "5", "5", "0", "0", "0"),
   sex = c(1, 2, 3, 1, 2, 1),
   avail = c(0, 1, 0, 1, 0, 1),
   affected = matrix(c(
       0, 1, 0, 1, 0, 1,
       1, 1, 1, 1, 1, 1
   ), ncol = 2),
   col_aff = c("aff1", "aff2"),
   missid = "0",
   rel_df = matrix(c(
      "1", "2", 2
   ), ncol = 3, byrow = TRUE),
)
data(sampleped)
Pedigree(sampleped)
Pedigree()
}
\seealso{
\code{\link[=Pedigree]{Pedigree()}}, \code{\link[=Ped]{Ped()}}, \code{\link[=Rel]{Rel()}}, \code{\link[=Scales]{Scales()}}
}
