% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{pedigree}
\alias{pedigree}
\title{Create a Pedigree object from a data.frame}
\usage{
pedigree(
  ped_df = data.frame(id = character(), dadid = character(), momid = character(), sex =
    numeric(), family = character()),
  rel_df = data.frame(id1 = character(), id2 = character(), code = numeric(), family =
    character()),
  cols_ren_ped = list(indId = "id", fatherId = "dadid", motherId = "momid", gender =
    "sex"),
  cols_ren_rel = list(indId1 = "id1", indId2 = "id2"),
  scales = list(fill = data.frame(order = numeric(), column_values = character(),
    column_mods = character(), mods = numeric(), labels = character(), affected =
    logical(), fill = character(), density = numeric(), angle = numeric()), border =
    data.frame(column = character(), mods = numeric(), labels = character(), border =
    character())),
  hints = list(order = NULL, spouse = NULL),
  normalize = TRUE,
  ...
)
}
\arguments{
\item{ped_df}{A data.frame with the individuals informations.}

\item{rel_df}{A data.frame with the special relationships between
individuals.}

\item{cols_ren_ped}{A named list with the columns to rename for the
pedigree dataframe.}

\item{cols_ren_rel}{A named list with the columns to rename for the
relationship matrix.}

\item{scales}{A list of two data.frame with the scales to use for the
affection status and the other one for the border color (e.g availability).}

\item{normalize}{A logical to know if the data should be normalised.}

\item{...}{Other arguments to pass to the function `generate_colors`.}
}
\value{
A Pedigree object.
}
\description{
This constructor help to create a \\code{Pedigree} object from
`data.frame`.
}
