% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{pedigree}
\alias{pedigree}
\title{Create a Pedigree object from a data.frame}
\usage{
pedigree(
  ped_df = data.frame(id = character(), dadid = character(), momid = character(), sex =
    numeric(), family = character()),
  rel_df = data.frame(id1 = character(), id2 = character(), code = numeric()),
  cols_ren_ped = list(indId = "id", fatherId = "dadid", motherId = "momid", gender =
    "sex"),
  scales = data.frame(column = character(), mods_labels = character(), fill =
    character(), border = character(), density = numeric(), angle = numeric()),
  normalize = TRUE
)
}
\arguments{
\item{ped_df}{A data.frame with the individuals informations.}

\item{rel_df}{A data.frame with the special relationships between
individuals.}

\item{cols_ren_ped}{A named list with the columns to rename.}

\item{scales}{A data.frame with the scales to use for the affection status.}

\item{normalize}{A logical to know if the data should be normalised.}
}
\value{
A Pedigree object.
}
\description{
This constructor help to create a \\code{Pedigree} object from
`data.frame`.
}
