% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pedigree.shrink.R
\name{plot.pedigree.shrink}
\alias{plot.pedigree.shrink}
\title{Plot shrinked pedigree}
\usage{
\method{plot}{pedigree.shrink}(x, bigped = FALSE, title = "", location = "topright", ...)
}
\arguments{
\item{x}{A pedigree.shrink object, which contains a pedigree object and
information about which subject was removed.}

\item{bigped}{Logical value indicating whether pedigree should be compacted
to fit in plotting region.  If T, then packed=F is used in pedigree() along
with smaller symbol sizes.}

\item{title}{Optional plot title}

\item{location}{The position argument for the legend command, which allows
coordinates (numerical vector in the form x, y) or, more conveniently,
options such as "topright", "right", "left", "bottomleft", etc., which
is useful for pedigrees that cover most of the plot region.}

\item{...}{Optional arguments to plot method}
}
\value{
Plot shrinked pedigree
}
\description{
plot pedigree.shrink object that is a shrunk pedigree object
}
\details{
Plot the pedigree object that is the trimmed pedigree.shrink object along
with colors based on availability and affection status.
}
\examples{

data(sample.ped)

fam2 <- sample.ped[sample.ped$ped == 2, ]
ped2 <- pedigree(
  fam2$id, fam2$father, fam2$mother, fam2$sex,
  fam2$affected, fam2$avail
)

shrink2 <- pedigree.shrink(ped2, avail = fam2$avail)
shrink2

plot(ped2)

plot.pedigree.shrink(shrink2, title = "Sample Pedigree 2")

}
\seealso{
`pedigree.shrink`
}
