% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{polyfun}
\alias{polyfun}
\title{Polygonal element}
\usage{
polyfun(nslice, coor)
}
\arguments{
\item{nslice}{Number of slices in the polygon}

\item{coor}{Element form which to generate the polygon
containing x and y coordinates and theta}
}
\value{
a list of x and y coordinates
}
\description{
Create a list of x and y coordinates for a polygon
with a given number of slices and a list of coordinates
for the polygon.
}
\examples{
polyfun(2, list(
     x = c(-0.5, -0.5, 0.5, 0.5),
     y = c(-0.5, 0.5, 0.5, -0.5),
     theta = -c(3, 5, 7, 9) * pi / 4
))
}
\keyword{Pedigree-plot}
\keyword{internal,}
