% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normData.R
\name{prefix_famid}
\alias{prefix_famid}
\title{Compute id with family id}
\usage{
prefix_famid(family_id, ind_id, missid = "0")
}
\arguments{
\item{family_id}{The family id}

\item{ind_id}{The individual id}

\item{missid}{The missing id}
}
\value{
The id
}
\description{
Compute id with family id if the family id available
}
