% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\name{prefix_famid}
\alias{prefix_famid}
\title{Compute id with family id}
\usage{
prefix_famid(famid, id, missid = NA_character_)
}
\arguments{
\item{famid}{A character vector with the family identifiers of the
individuals. If provide, will be aggregated to the individuals
identifiers separated by an underscore.}

\item{missid}{A character vector with the missing values identifiers.
All the id, dadid and momid corresponding to those values will be set
to \code{NA_character_}.}
}
\value{
The id with the family id merged
}
\description{
Compute id with family id if the family id available
}
\keyword{internal}
