% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_data.R
\name{prefix_famid}
\alias{prefix_famid}
\title{Compute id with family id}
\usage{
prefix_famid(family_id, ind_id, missid = "0")
}
\arguments{
\item{family_id}{The family id}

\item{ind_id}{The individual id}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree
(i.e. \code{dadid} and \code{momid} equal to the value of this variable).
The default for \code{missid} is \code{"0"}.}
}
\value{
The id with the family id merged
}
\description{
Compute id with family id if the family id available
}
\keyword{internal}
