% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{summary,Rel-method}
\alias{summary,Rel-method}
\alias{show,Rel-method}
\alias{as.list,Rel-method}
\alias{as.data.frame,Rel-method}
\title{Summary function of Rel object}
\usage{
\S4method{summary}{Rel}(object)

\S4method{show}{Rel}(object)

\S4method{as.list}{Rel}(x)

\S4method{as.data.frame}{Rel}(x)
}
\arguments{
\item{object}{A Rel object.}

\item{from}{A Rel object.}
}
\value{
A character vector with the summary of the object.

The Rel object with the relationship informations.

A list with the relationship informations.

A data.frame with the relationship informations.
}
\description{
Compute the summary of a Rel object

Convert the Rel object to a data.frame
and print it with its summary.

Convert a Rel object to a list

Convert a Rel object to a data.frame
}
