% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{set_plot_area}
\alias{set_plot_area}
\title{Set plotting area}
\usage{
set_plot_area(cex, id, maxlev, xrange, symbolsize, ...)
}
\arguments{
\item{cex}{Character expansion of the text}

\item{id}{A character vector with the identifiers of each individuals}

\item{maxlev}{Maximum level}

\item{xrange}{Range of x values}

\item{symbolsize}{Size of the symbols}

\item{...}{Other arguments passed to \code{\link[=par]{par()}}}
}
\value{
List of user coordinates, old par, box width, box height,
label height and leg height
}
\description{
Set plotting area
}
\keyword{Pedigree-plot}
\keyword{internal,}
