% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_hint.R
\name{shift}
\alias{shift}
\title{Shift set of siblings to the left or right}
\usage{
shift(id, sibs, goleft, hint, twinrel, twinset)
}
\arguments{
\item{id}{The id of the subject to be shifted}

\item{sibs}{The ids of the siblings}

\item{goleft}{If \code{TRUE}, shift to the left, otherwise to the right}

\item{hint}{The current hint vector}

\item{twinrel}{The twin relationship matrix}

\item{twinset}{The twinset vector}
}
\value{
The updated hint vector
}
\description{
Shift set of siblings to the left or right
}
\details{
This routine is used by \code{auto_hint()}.
It shifts a set of siblings to the left or right, so that the
marriage is on the edge of the set of siblings, closest to the spouse.
It also shifts the subject himself, so that he is on the edge of the
set of siblings, closest to the spouse.
It also shifts the monozygotic twins, if any, so that they are
together within the set of twins.
}
\seealso{
\code{\link[=auto_hint]{auto_hint()}}
}
\keyword{auto_hint}
\keyword{internal,}
