% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{subregion}
\alias{subregion}
\title{Routine to subset a pedigree}
\usage{
subregion(plist, subreg)
}
\arguments{
\item{subreg}{4-element vector for (min x, max x, min depth, max depth),
used to edit away portions of the plot coordinates returned by
\code{\link[=align]{align()}}. This is useful for zooming in on a particular region of the
pedigree.}
}
\value{
a pedigree list
}
\description{
Routine to subset a pedigree
}
