% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{subregion}
\alias{subregion}
\title{Routine to subset a pedigree}
\usage{
subregion(plist, subreg)
}
\arguments{
\item{subreg}{a vector of length 4 giving the start and end positions
of the subregion and the start and end depths}
}
\value{
a pedigree list
}
\description{
Routine to subset a pedigree
}
