% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim a pedigree}
\usage{
trim(ped, id_rm, missid = "0")
}
\arguments{
\item{ped}{A pedigree object}

\item{id_rm}{Vector of ids to remove}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree (i.e. \code{dadid} and \code{momid} equal to
the value of this variable).  The default for \code{missid} is \code{"0"}.}
}
\value{
A Pedigree object with the subjects removed
}
\description{
Carries out the removal of the subjects identified from a pedigree object.
}
