% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_inds.R
\name{useful_inds}
\alias{useful_inds}
\title{Compute the usefulness of individuals}
\usage{
useful_inds(df, informative = "AvAf")
}
\arguments{
\item{df}{Dataframe with individuals to select}

\item{informative}{Informative individuals selection can take 3 values:
"AvAf" (available and affected),
"AvOrAf" (available or affected),
"Av" (available only),
"Af" (affected only),
"All" (all individuals)
or a numeric vector of individuals id
or a boolean}
}
\description{
Check for usefulness of individuals
}
\details{
Check for the informativeness of the individuals based on the
informative parameter given, the number of children and the usefulness
of their parents. A `useful` column is added to the dataframe with the
usefulness of the individual. This boolean is hereditary.
}
