% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_inds.R
\docType{methods}
\name{useful_inds}
\alias{useful_inds}
\alias{useful_inds,character-method}
\alias{useful_inds,character}
\alias{useful_inds,Pedigree-method}
\alias{useful_inds,Pedigree}
\alias{useful_inds,Ped-method}
\title{Compute the usefulness of individuals}
\usage{
useful_inds(obj, ...)

\S4method{useful_inds}{character}(
  obj,
  dadid,
  momid,
  avail,
  affected,
  num_child_tot,
  informative = "AvAf",
  keep_infos = FALSE,
  missid = NA_character_
)

\S4method{useful_inds}{Pedigree}(obj, informative = "AvAf", keep_infos = FALSE, reset = FALSE)

\S4method{useful_inds}{Ped}(obj, informative = "AvAf", keep_infos = FALSE, reset = FALSE)
}
\arguments{
\item{obj}{A character vector with the id of the individuals or a
\code{data.frame} with all the informations in corresponding columns.}

\item{...}{Additional arguments passed to methods}

\item{avail}{A logical vector with the availability status of the
individuals
(i.e. \code{FALSE} = not available, \code{TRUE} = available, \code{NA} = unknown).}

\item{affected}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = unaffected, \code{TRUE} = affected, \code{NA} = unknown).}

\item{num_child_tot}{A numeric vector of the number of children of each
individuals}

\item{informative}{Informative individuals selection can take 5 values:
\itemize{
\item 'AvAf' (available and affected),
\item 'AvOrAf' (available or affected),
\item 'Av' (available only),
\item 'Af' (affected only),
\item 'All' (all individuals)
\item A numeric/character vector of individuals id
\item A boolean
}}

\item{keep_infos}{Boolean to indicate if individuals with unknown status
but available or reverse should be kept}

\item{missid}{A character vector with the missing values identifiers.
All the id, dadid and momid corresponding to those values will be set
to \code{NA_character_}.}

\item{reset}{Boolean to indicate if the \code{useful} column should be reset}
}
\value{
\subsection{When obj is a vector}{

A vector of useful individuals identifiers
}

\subsection{When obj is a Pedigree}{

The Pedigree object with a new column named 'useful' containing 1 for
useful individuals and 0 otherwise.
}
}
\description{
Check for usefulness of individuals
}
\details{
Check for the informativeness of the individuals based on the
informative parameter given, the number of children and the usefulness
of their parents. A \code{useful} column is added to the dataframe with the
usefulness of the individual. This boolean is hereditary.
}
\examples{
data(sampleped)
ped1 <- Pedigree(sampleped[sampleped$famid == "1",])
ped1 <- num_child(ped1)
useful_inds(ped1, informative = "AvAf")$ped
}
