% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_inds.R
\docType{methods}
\name{useful_inds}
\alias{useful_inds}
\alias{useful_inds,character-method}
\alias{useful_inds,character}
\alias{useful_inds,Pedigree-method}
\alias{useful_inds,Pedigree}
\title{Compute the usefulness of individuals}
\usage{
useful_inds(obj, ...)

\S4method{useful_inds}{character}(
  obj,
  dadid,
  momid,
  avail,
  affected,
  num_child_tot,
  informative = "AvAf",
  keep_infos = FALSE,
  missid = "0"
)

\S4method{useful_inds}{Pedigree}(
  obj,
  informative = "AvAf",
  keep_infos = FALSE,
  missid = "0",
  reset = FALSE
)
}
\arguments{
\item{obj}{A pedigree object or a vector of subject identifiers.}

\item{...}{Additional arguments passed to methods}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{avail}{A numeric vector of availability status of each individual
(e.g., genotyped). The values are:
\itemize{
\item \code{0}  : unavailable
\item \code{1}  : available
\item \code{NA} : availability not known
}}

\item{affected}{A numeric vector of affection status of each individual
(e.g., genotyped). The values are:
\itemize{
\item \code{0}  : unaffected
\item \code{1}  : affected
\item \code{NA} : affection status not known
}}

\item{num_child_tot}{A numeric vector of the number of children of each
individuals}

\item{informative}{Informative individuals selection can take 5 values:
\itemize{
\item 'AvAf' (available and affected),
\item 'AvOrAf' (available or affected),
\item 'Av' (available only),
\item 'Af' (affected only),
\item 'All' (all individuals)
\item A numeric/character vector of individuals id
\item A boolean
}}

\item{keep_infos}{Boolean to indicate if individuals with unknown status
but available or reverse should be kept}

\item{missid}{The missing identifier value. Founders are the individuals with
no father and no mother in the pedigree (i.e. \code{dadid} and \code{momid} equal to
the value of this variable).  The default for \code{missid} is \code{"0"}.}

\item{reset}{Boolean to indicate if the \code{useful} column should be reset}
}
\value{
\subsection{When obj is a vector}{

A vector of useful individuals identifiers
}

\subsection{When obj is a Pedigree}{

The Pedigree object with a new column named 'useful' containing 1 for
useful individuals and 0 otherwise.
}
}
\description{
Check for usefulness of individuals
}
\details{
Check for the informativeness of the individuals based on the
informative parameter given, the number of children and the usefulness
of their parents. A \code{useful} column is added to the dataframe with the
usefulness of the individual. This boolean is hereditary.
}
