% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_inds.R
\name{useful_inds}
\alias{useful_inds}
\alias{useful_inds,character-method}
\alias{useful_inds,Pedigree-method}
\alias{useful_inds,Ped-method}
\title{Usefulness of individuals}
\usage{
\S4method{useful_inds}{character}(
  obj,
  dadid,
  momid,
  avail,
  affected,
  num_child_tot,
  informative = "AvAf",
  keep_infos = FALSE
)

\S4method{useful_inds}{Pedigree}(obj, informative = "AvAf", keep_infos = FALSE, reset = FALSE)

\S4method{useful_inds}{Ped}(obj, informative = "AvAf", keep_infos = FALSE, reset = FALSE)
}
\arguments{
\item{obj}{A character vector with the id of the individuals or a
\code{data.frame} with all the informations in corresponding columns.}

\item{dadid}{A vector containing for each subject, the identifiers of the
biologicals fathers.}

\item{momid}{A vector containing for each subject, the identifiers of the
biologicals mothers.}

\item{avail}{A logical vector with the availability status of the
individuals
(i.e. \code{FALSE} = not available, \code{TRUE} = available, \code{NA} = unknown).}

\item{affected}{A logical vector with the affection status of the
individuals
(i.e. \code{FALSE} = unaffected, \code{TRUE} = affected, \code{NA} = unknown).}

\item{num_child_tot}{A numeric vector of the number of children of each
individuals}

\item{informative}{Informative individuals selection can take 5 values:
\itemize{
\item 'AvAf' (available and affected),
\item 'AvOrAf' (available or affected),
\item 'Av' (available only),
\item 'Af' (affected only),
\item 'All' (all individuals)
\item A numeric/character vector of individuals id
\item A boolean
}}

\item{keep_infos}{Boolean to indicate if individuals with unknown status
but available or reverse should be kept}

\item{reset}{Boolean to indicate if the \code{useful} column should be reset}
}
\value{
\subsection{When obj is a vector}{

A vector of useful individuals identifiers
}

\subsection{When obj is a Pedigree or Ped object}{

The Pedigree or Ped object with the slot 'useful' containing \code{TRUE} for
useful individuals and \code{FALSE} otherwise.
}
}
\description{
Compute the usefulness of individuals
}
\details{
Check for the informativeness of the individuals based on the
informative parameter given, the number of children and the usefulness
of their parents. A \code{useful} slot is added to the Ped object with the
usefulness of the individual. This boolean is hereditary.
}
\examples{

data(sampleped)
ped1 <- Pedigree(sampleped[sampleped$famid == "1",])
ped(useful_inds(ped1, informative = "AvAf"))
}
\keyword{shrink}
